% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{GUI app for interactive whale-strike simulations}
\usage{
app(mode = "simple", options = list(height = 500))
}
\arguments{
\item{mode}{character value specifying the style to use.  Only
the value \code{"simple"} is permitted at present. This yields a
3-panel plot, constructed by \code{\link{plot.strike}},
called with default arguments.}

\item{options}{list containing options that are provided
to \code{\link[shiny]{shinyApp}}, which creates the GUI app.}
}
\description{
GUI app for interactive whale-strike simulations
}
\author{
Dan Kelley
}
