% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{areaFromLength}
\alias{areaFromLength}
\title{Whale projected area, as function of length}
\usage{
areaFromLength(L, view = "side")
}
\arguments{
\item{L}{length in m}

\item{view}{character string indicating the viewpoint; only
\code{"side"} is permitted at present.}
}
\description{
Whale projected area, as function of length
}
