% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{blubberForce}
\alias{blubberForce}
\title{Blubber force}
\usage{
blubberForce(x, X, parms)
}
\arguments{
\item{x}{Boat position [m]}

\item{X}{Whale position [m]}

\item{parms}{Parameters of the simulation (as provided to \code{\link{simulate}}).}
}
\value{
Compression-resisting force [N]
}
\description{
Blubber force
}
