% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{contactArea}
\alias{contactArea}
\title{Contact area}
\usage{
contactArea(x, X, parms)
}
\arguments{
\item{x}{Boat position [m]}

\item{X}{Whale position [m]}

\item{parms}{Parameters of the simulation (as provided to \code{\link{simulate}}).}
}
\value{
Contact area [m^2].
}
\description{
Area of contact between vessel and whale skin
}
