% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{derivative}
\alias{derivative}
\title{Calculate derivative using first difference}
\usage{
derivative(var, t)
}
\arguments{
\item{var}{variable.}

\item{t}{time in seconds.}
}
\value{
Derivative estimated by using \code{\link[=diff]{diff()}} on both \code{var}
and \code{time}.
}
\description{
The derivative is estimated as the ratio of the first-difference of \code{var}
divided by the first-difference of \code{time}.  To make the results
have the same length as \code{time}, the final result is appended at
the end.
}
\author{
Dan Kelley
}
