% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{dynamics}
\alias{dynamics}
\title{Dynamical law}
\usage{
dynamics(t, y, parms)
}
\arguments{
\item{t}{time [s].}

\item{y}{model state, a vector containing boat position \code{X} [m],
boat speed \code{V} [m/s], whale position \code{x} [m],
and whale speed \code{v} [m/s].}

\item{parms}{model parameters.}
}
\description{
Dynamical law
}
