% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{lengthFromMass}
\alias{lengthFromMass}
\title{Compute whale length from mass}
\usage{
lengthFromMass(M, model = "fortune2012atlantic")
}
\arguments{
\item{M}{Whale mass [kg].}

\item{model}{Character string specifying the model, with permitted
values \code{"moore2005"}, \code{"fortune2012atlantic"} and
for \code{\link{massFromLength}} for the details of these
formulations.}
}
\value{
Whale length [m].
}
\description{
This works by inverting \code{\link{massFromLength}} using
\code{\link{uniroot}}.
}
