% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{lethalityIndexFromStress}
\alias{lethalityIndexFromStress}
\title{Compute lethality index, based on compression stress}
\usage{
lethalityIndexFromStress(stress)
}
\arguments{
\item{stress}{numerical value or vector, giving whale compression stress in Pascals.}
}
\value{
threat of injury (in range 0 to 1)
}
\description{
The model used for this is the logistic model, fitting observed injury/lethality
statistics to the base-10 logarithm of the maximum compression stress during
a simulated impact event.
}
\examples{
lethalityIndexFromStress(parameters()$logistic$tau50) # approx. 0.5

}
\seealso{
Other functions dealing with Whale Lethality index: \code{\link{stressFromLethalityIndex}}
}
\author{
Dan Kelley
}
\concept{functions dealing with Whale Lethality index}
