% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{massFromLength}
\alias{massFromLength}
\title{Whale mass inferred from length}
\usage{
massFromLength(L, model = 1)
}
\arguments{
\item{L}{Whale length in m.}

\item{model}{A number indicating the model (see \dQuote{Details}).}
}
\value{
Mass in kg.
}
\description{
Calculate an estimate of the mass of a whale, based on its length.
}
\details{
The permitted models are as follows.
\itemize{
\item \code{model=1} yields
\eqn{242.988 * exp(0.4 * length)}{242.988 * exp(0.4 * length)},
which (apart from a unit change) is a regression equation
shown above Figure 1d of Moore et al. (2005) for right whales.
}
}
\examples{
library(whalestrike)
L <- seq(5, 15, length.out=100)
plot(L, massFromLength(L)/1000, type='l',
     xlab="Right-whale Length [m]", ylab="Mass [tonne]")
grid()
}
\references{
Moore, M.J., A.R. Knowlton, S.D. Kraus, W.A. McLellan, and R.K. Bonde.
“Morphometry, Gross Morphology and Available Histopathology in North Atlantic
Right Whale (Eubalaena Glacialis) Mortalities (1970 to 2002).” Journal of
Cetacean Research and Management 6, no. 3 (2005): 199–214.
}
