% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lethality_index.R
\name{maximumLethalityIndex}
\alias{maximumLethalityIndex}
\title{Find maximum Lethality Index during a strike}
\usage{
maximumLethalityIndex(strike)
}
\arguments{
\item{strike}{the value returned by a call to strike.}
}
\value{
The maximum value of the Lethality Index that is
involved in the simulation of the ship-whale collision event.
This is a unitless number; see Kelley et al. (2021).
}
\description{
This works by finding the maximum Lethality Index encountered during
a simulation created by calling \code{\link[=strike]{strike()}}, and so it is important to
use a detailed setting for the output times. In the example, the
results are reported every 0.7/200 seconds (i.e. 3.5 milliseconds),
which is likely sufficient (see the example, where a plot is
used for this assessment).
}
\examples{
library(whalestrike)
t <- seq(0, 0.7, length.out = 200)
state <- list(xs = -2, vs = knot2mps(10), xw = 0, vw = 0)
parms <- parameters()
s <- strike(t, state, parms)
# Compute the desired value and (for context) show it on a plot
maximumLethalityIndex(s)
# For context, this is how this can be done "by hand"
max(lethalityIndexFromStress(s[["WCF"]][["stress"]]))
# Show the maximum on a plot (see also the plot title)
plot(s, which = "lethality index")
abline(h=maximumLethalityIndex(s), col=2)

}
\references{
Kelley, Dan E., James P. Vlasic, and Sean W. Brillant. "Assessing the
Lethality of Ship Strikes on Whales Using Simple Biophysical Models." Marine
Mammal Science 37, no. 1 (January 2021): 251–67.
}
\seealso{
Other functions dealing with Whale Lethality index: 
\code{\link{lethalityIndexFromStress}()},
\code{\link{stressFromLethalityIndex}()}
}
\author{
Dan Kelley, wrapping code provided by Alexandra Mayette
}
\concept{functions dealing with Whale Lethality index}
