% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mps2knot}
\alias{mps2knot}
\title{Convert a speed in m/s to a speed in knots}
\usage{
mps2knot(mps)
}
\arguments{
\item{mps}{Speed in metres per second.}
}
\value{
Speed in knots.
}
\description{
This is done by dividing by the factor 1.852e3/3600,
See also \code{\link[=knot2mps]{knot2mps()}}, which is the inverse of this function.
}
\examples{
library(whalestrike)
mps <- seq(0, 10)
plot(mps, mps2knot(mps), xlab = "Speed [m/s]", ylab = "Speed [knots]", type = "l")

}
\seealso{
Other functions dealing with units: 
\code{\link{knot2mps}()}
}
\author{
Dan Kelley
}
\concept{functions dealing with units}
