% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{parameters}
\alias{parameters}
\title{Control parameters for whale strike simulation}
\usage{
parameters(ms, Ss, B = 3, D = 1.5, mw, lw, Sw, delta = 0.01, Es,
  theta = 45, beta = 0.3, Eb = 6e+05, alpha = 1, Ea = 4e+05,
  CDs = 0.002, CDw = 0.003)
}
\arguments{
\item{ms}{Ship mass [kg].}

\item{Ss}{Ship wetted area [m^2]. This, together with \code{CDs}, is used by
used by \code{\link{shipWaterForce}} to estimate ship drag force.}

\item{B}{Ship impact horizontal extent [m]; defaults to 2m if not specified.}

\item{D}{Ship impact vertical extent [m]; defaults to 1.5m if not specified.}

\item{mw}{Whale mass [kg]. If not supplied, \code{\link{whaleMassFromLength}} is used
to calculate this, given \code{lw}, but if neither \code{mw} nor \code{ml} is provided,
an error is reported.}

\item{lw}{Whale length [m]. If not supplied, \code{\link{whaleLengthFromMass}} is used
to calculate this, given \code{lm}, but if neither \code{mw} nor \code{ml} is provided,
an error is reported. The length is used by \code{\link{whaleAreaFromLength}} to
calculate area, which is needed for the water drag calculation done by
\code{\link{whaleWaterForce}}.}

\item{Sw}{Whale surface area [m^2]. This, together with \code{CDw}, is used by
used by \code{\link{whaleWaterForce}} to estimate whale drag force.}

\item{delta}{Whale skin thickness [m]. Defaults to 0.01 m, if not supplied.}

\item{Es}{Whale skin elastic modulus [Pa]. If not provided, defaults to 20e6 Pa,
a rounded value of the 19.56+-4.03MPa estimate for seal skin, provided in Table 3
of Grear et al. (2017).}

\item{theta}{Whale skin deformation angle [deg]; defaults to 45deg if not supplied.}

\item{beta}{Whale blubber thickness [m]; defaults to 0.3m if not supplied.}

\item{Eb}{Elastic modulus of blubber [Pa]; defaults to 6e5 Pa (the value
suggested Raymond (2007 fig 37), rounded to 1 digit), if not supplied.}

\item{alpha}{Thickness of interior region [m]. This defaults to 1m, an 
estimate of the radius of soft tissue below the blubber. For bone,
a value in the centimeter range might be reasonable.}

\item{Ea}{Elastic modulus of interior region [Pa]; defaults to 4e5 Pa,
for soft tissue as stated by Raymond (2007) page 36 (he actually states
425294 Pa). For bone, the value 9e8 Pa might be reasonable (see Raymond 2007
Table 2.3, which lists the elastic modulus of cortical bone as 854.2 MPa).}

\item{CDs}{Drag coefficient for ship [dimensionless],
used by \code{\link{shipWaterForce}} to estimate ship drag force. Defaults
to 2.5e-3, using Figure 4 of Manen and van Oossanen (1988), assuming
Reynolds number of 5e7, computed from speed 5m/s, lengthscale 10m
and viscosity 1e-6 m^2/s. The drag force is computed with
\code{\link{shipWaterForce}}.}

\item{CDw}{Drag coefficient for whale [dimensionless],
used by \code{\link{whaleWaterForce}} to estimate whale drag force.
Defaults to 3.0e-3, for Reynolds number 2e7, computed from speed
2 m/s, lengthscale 5m (between radius and length) and
viscosity 1e-6 m^2/s.  The drag force is computed with
\code{\link{whaleWaterForce}}.}
}
\value{
A named list holding the parameters, with defaults and alternatives reconciled
according to the system described above.
}
\description{
Assembles control parameters into a list suitable for passing to \code{\link{strike}}
and the functions that it calls.
}
\references{
See \link{whalestrike} for a list of references.
}
