% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{pin}
\alias{pin}
\title{Pin numerical values between stated limits}
\usage{
pin(x, lower = NULL, upper = NULL)
}
\arguments{
\item{x}{Vector or matrix of numerical values}

\item{lower}{Numerical values of minimum value allowed; set to \code{NULL}
to avoid trimming the lower limit.}

\item{upper}{As for \code{lower}, but for the upper limit.}
}
\value{
Copy of \code{x}, with any value that exceeds \code{lim} having
been replaced by \code{lim}.
}
\description{
Pin numerical values between stated limits
}
\author{
Dan Kelley
}
