% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{plot.whalestrike}
\alias{plot.whalestrike}
\title{Plot a whalestrike object}
\usage{
\method{plot}{whalestrike}(x, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{whalestrike}}

\item{...}{Other arguments (ignored).}
}
\description{
Plot a whalestrike object
}
