% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\docType{data}
\name{raymond2007}
\alias{raymond2007}
\title{Whale blubber stress-strain relationship}
\description{
This is a data frame with elements \code{strain} and \code{stess},
found by digitizing (accurate to perhaps 1 percent) the curve shown in Figure 2.13
of Raymond (2007).

Raymond, J. J. “Development of a Numerical Model to Predict Impact Forces on a
North Atlantic Right Whale during Collision with a Vessel.” University of New
Hampshire, 2007. \url{https://scholars.unh.edu/thesis/309}.
}
