% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{shipAreaFromMass}
\alias{shipAreaFromMass}
\title{Compute ship wetted area from mass}
\usage{
shipAreaFromMass(ms)
}
\description{
The calculation is used by \code{\link{shipWaterForce}}. The method
is to scale up the results for a known Cape
Islander ship, of displacement 20.46 tonnes, length 11.73m,
beam 4.63m, and draft 1.58m, on the assumption that the wetted area
is length*(2*draft+beam). This is a very crude model, but it is not
likely the weakest link in the modelling of ship dynamics.
}
