% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship.R
\name{shipLength}
\alias{shipLength}
\title{Nominal ship length in m}
\usage{
shipLength(type = NULL)
}
\arguments{
\item{type}{either (1) a string identifying the ship type, in
which case the average overall length of the named vessel is
returned, or (2) NULL, in which case a data frame containing
type and length is returned.}
}
\value{
\code{shipLength} returns ship length in m, as defined in Mayette
and Brillant (2026).
}
\description{
This is based on the "Average LOA in m" column in Table 1
of Mayette and Brillant (2026).
}
\examples{
library(whalestrike)
# An individual length
shipLength("Fishing")
# A table of lengths
shipLength()

}
\references{
Mayette, Alexandra, and Sean W. Brillant. "A Regression-Based Method
to Estimate Vessel Mass for Use in Whale-Ship Strike Risk Models."
PloS One 21, no. 1 (2026): e0339760.
https://doi.org/10.1371/journal.pone.0339760.
}
\seealso{
Other functions relating to ship characteristics: 
\code{\link{shipAreaFromMass}()},
\code{\link{shipMassFromLength}()},
\code{\link{shipWaterForce}()}
}
\author{
Dan Kelley, with help from Alexandra Mayette
}
\concept{functions relating to ship characteristics}
