% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship.R
\name{shipMassFromLength}
\alias{shipMassFromLength}
\title{Compute ship displacement from vessel type and length}
\usage{
shipMassFromLength(type, L)
}
\arguments{
\item{type}{a string identifying the ship type, from the list
given in \sQuote{Details}.}

\item{L}{vessel length in metres.}
}
\value{
\code{shipMassFromLength} returns ship displacement mass
(in kg), according to Mayette and Brillant (2026) Table 3.
}
\description{
This is done using formulae in Table 3 of Mayette and
Brillant (2026).
}
\details{
The formulae used are as follows.\tabular{lr}{
   \code{type} \tab Formula \cr
   "Bulk Carrier" \tab 5.64 * L^3.06 \cr
   "Container Ship" \tab 86.40 * L^2.46 \cr
   "Cruise" \tab 97.51 * L^2.28 \cr
   "Ferry" \tab 25.15 * L^2.62 \cr
   "Fishing" \tab 0.71 * L^3.79 \cr
   "Government/Research" \tab 2.95 * L^3.22 \cr
   "Other" \tab 2.64 * L^3.35 \cr
   "Passenger" \tab 4.32 * L^3.08 \cr
   "Pleasure Craft" \tab 34.47 * L^2.68 \cr
   "Sailing" \tab 1.23 * L^3.53 \cr
   "Tanker" \tab 7.25 * L^3.03 \cr
   "Tug" \tab 104.48 * L^2.51 \cr
}
}
\references{
\itemize{
\item Alexandra Mayette, Sean W. Brillant. "A regression-based method
to estimate vessel mass for use in whale-ship strike risk models."
PLoS One 21(2) e0339760. \doi{10.1371/journal.pone.0339760}
}
}
\author{
Dan Kelley
}
