% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forces.R
\name{shipWaterForce}
\alias{shipWaterForce}
\title{Ship water force}
\usage{
shipWaterForce(vs, parms)
}
\arguments{
\item{vs}{Ship speed in m/s. (Consider using \code{\link[=knot2mps]{knot2mps()}} if you
prefer to think of speeds in knots.)}

\item{parms}{A named list holding model parameters, created by
\code{\link[=parameters]{parameters()}}.}
}
\value{
Water drag force (N).
}
\description{
Compute the retarding force of water on the ship, based on a drag law
\eqn{(1/2)*rho*Cs*A*vs^2}{(1/2)*rho*Cs*A*vs^2}
where \code{rho} is water density taken to be 1024 (kg/m^3), \code{Cs} is drag coefficient
stored in \code{parms} and \code{A} is area, also stored in \verb{parms, and }vs` is
the ship speed (m/s).
}
\author{
Dan Kelley
}
