% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\docType{data}
\name{sol20200708}
\alias{sol20200708}
\title{Reference strike() solution}
\description{
This was produced with the package as it existed on 2020-jul-8,
prior to the publication of Kelley et al. (2020).  It is used
in testing, to ensure that the package does not inadvertently
change in its predictions.
}
\examples{
data(sol20200708)
par(mfrow=c(1, 3))
plot(sol20200708)

}
\references{
Kelley, Dan E., James P. Vlasic, and Sean W. Brillant. “Assessing the Lethality
of Ship Strikes on Whales Using Simple Biophysical Models.” Marine Mammal
Science, October 12, 2020, mms.12745. (https://doi.org/10.1111/mms.12745).
}
