% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dan.R, R/parameters.R
\name{speciesParameters}
\alias{speciesParameters}
\title{Look up some parameeters by species}
\usage{
speciesParameters(species)

speciesParameters(species)
}
\arguments{
\item{species}{a string giving the whale species.  This must be one of
the following:
\code{"Blue"}, \code{"Fin"}, \code{"Gray"}, \code{"Humpback"}, \code{"Minke"}, \code{"N. Atl. Right"},
\code{"Sei"} or \code{"Sperm"}. FIXME: add in the two species not in AM's list, with a note
about what is done.}
}
\value{
a named vector holding \code{length} in metres, along with
the thicknesses (in centimetres) of \code{bone}, \code{sublayer}, \code{blubber}
and \code{skin}.
}
\description{
This is done by consulting the built-in \code{whale_measurements} dataset.

This is done by consulting the built-in \code{whale_measurements} dataset. The purpose
is to simplify calls to \code{\link[=parameters]{parameters()}}. FIXME: mention about those 2 guessed-at
species.
}
\details{
# Arguments
}
\author{
Dan Kelley, using data provided by Alexandra Mayette.
}
