% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forces.R
\name{stressFromStrainFunction}
\alias{stressFromStrainFunction}
\title{Create a function for stress in laminated layers}
\usage{
stressFromStrainFunction(l, a, b, N = 1000)
}
\arguments{
\item{l}{vector of layer thicknesses}

\item{a}{vector of multipliers}

\item{b}{vector of e-fold parameters}

\item{N}{integer specifying how many segments to use in the spline}
}
\value{
A piecewise-linear function, created with \code{\link[=approxfun]{approxfun()}},
that returns stress as a function of total strain of the
system of compressing layers. For the purposes of the whale-strike
analysis, the strain should be between 0 and 1, i.e. there is
no notion of compressing blubber, etc. to negative thickness.
}
\description{
Denoting unforced layer thickness in the \eqn{i} layer as
\eqn{l_i} and strain there as \eqn{\epsilon_i=\Delta l_i/l_i},
we may write the stress-strain relationship as
\deqn{\sigma = a_i*(exp(b_i*\epsilon_i)-1)}
for each layer, where it is assumed that stress
\eqn{\sigma} is equal across layers.
Inverting this yields
\deqn{\epsilon_i= ln(1 + \sigma/a_i)/b_i}
where \eqn{ln} is the natural logarithm.  Therefore,
the change \eqn{\Delta L} in the total thickness \eqn{L=\sum l_i}
may be written
\deqn{0 = \Delta L - \sum((l_i/b_i) ln(1+\sigma/a_i))}.
Note that zero-thickness layers are removed from the calculation,
to avoid spurious forces.
}
\details{
This expression is not easily inverted to get
\eqn{\sigma} in terms of \eqn{\Delta L}
but it may be solved
easily for particular numerical values, using \code{\link[=uniroot]{uniroot()}}.

This is done for a sequence of \code{N} values of strain \eqn{\epsilon}
that range from 0 to 1. Then \code{\link[=approxfun]{approxfun()}} is used to create
a piecewise-linear representation of the relationship between \eqn{\sigma} and \eqn{\Delta L},
which becomes the return value of the present function.
(The purpose of using a piecewise-linear representation to reduce
computation time.)
}
\examples{
library(whalestrike)
# Set blubber parameters for each layer, to see if
# we recover the raymond2007 data.
param <- parameters(a = rep(1.64e5, 4), b = rep(2.47, 4))
x <- seq(0, 0.5, length.out = 100)
y <- param$stressFromStrain(x)
plot(x, y, type = "l", lwd = 4, col = "gray")
data("raymond2007")
points(raymond2007$strain, raymond2007$stress, col = 2)

}
\seealso{
Other functions relating to whale characteristics: 
\code{\link{whaleCompressionForce}()},
\code{\link{whaleLengthFromMass}()},
\code{\link{whaleMassFromLength}()},
\code{\link{whaleSkinForce}()},
\code{\link{whaleWaterForce}()}

Other functions relating to forces: 
\code{\link{shipWaterForce}()},
\code{\link{whaleCompressionForce}()},
\code{\link{whaleSkinForce}()},
\code{\link{whaleWaterForce}()}
}
\author{
Dan Kelley
}
\concept{functions relating to forces}
\concept{functions relating to whale characteristics}
