% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{stressFromThreat}
\alias{stressFromThreat}
\title{Compute stress, based on threat of injury, based on logistic model}
\usage{
stressFromThreat(injury)
}
\arguments{
\item{injury}{numerical value or vector, giving threat of injury (in range 0 to 1).}
}
\value{
tau whale compression stress in Pascals.
}
\description{
Compute stress, based on threat of injury, based on logistic model
}
\examples{
expect_equal(stressFromThreat(0.5), parameters()$logistic$tau50, tolerance=0.01)

}
\seealso{
Other functions dealing with threat index: \code{\link{threatFromStress}}
}
\author{
Dan Kelley
}
\concept{functions dealing with threat index}
