% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{summarize}
\alias{summarize}
\title{Summarize the parameters of a simulation, and its results}
\usage{
summarize(object, style = "text")
}
\arguments{
\item{object}{An object inheriting from class \link{strike}}

\item{style}{Character value indicating the representation to be used.
If \code{style} is \code{"text"}, then the output is in a textual format. If it
is \code{"latex"}, then the output is in latex format.}
}
\description{
Summarize the parameters of a simulation, and its results
}
\references{
See \code{\link[=whalestrike]{whalestrike()}} for a list of references.
}
\author{
Dan Kelley
}
