% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{threatFromStress}
\alias{threatFromStress}
\title{Compute threat of injury from stress, based on logistic model}
\usage{
threatFromStress(tau)
}
\arguments{
\item{tau}{numerical value or vector, giving whale compression stress in Pascals.}
}
\value{
threat of injury (in range 0 to 1)
}
\description{
Compute threat of injury from stress, based on logistic model
}
\examples{
expect_equal(threatFromStress(parameters()$logistic$tau50), 0.5, tolerance=0.01)

}
\seealso{
Other functions dealing with threat index: \code{\link{stressFromThreat}}
}
\author{
Dan Kelley
}
\concept{functions dealing with threat index}
