% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{trimBelow}
\alias{trimBelow}
\title{Trim values to a stated upper limit}
\usage{
trimBelow(x, lim)
}
\arguments{
\item{x}{Vector or matrix of numerical values}

\item{lim}{Numerical value of maximum allowed}
}
\value{
Copy of \code{x}, with any value that exceeds \code{lim} having
been replaced by \code{lim}.
}
\description{
Trim values to a stated upper limit
}
