% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{ui}
\alias{ui}
\title{User interface for app}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
ui
}
\description{
User interface for app
}
\keyword{datasets}
