% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{updateParameters}
\alias{updateParameters}
\title{Update parameters}
\usage{
updateParameters(
  original,
  ms,
  Ss,
  Ly,
  Lz,
  species,
  lw,
  mw,
  Sw,
  l,
  a,
  b,
  s,
  theta,
  Cs,
  Cw,
  logistic
)
}
\arguments{
\item{original}{An object of class \code{"parameters"}, as created by \code{\link[=parameters]{parameters()}}
and perhaps later altered by previous calls to \code{updateParameters()}.}

\item{ms}{Ship mass (kg).}

\item{Ss}{Ship wetted area (m^2). This, together with \code{Cs}, is used by
used by \code{\link[=shipWaterForce]{shipWaterForce()}} to estimate ship drag force. If \code{Ss}
is not given, then an estimate is made by calling \code{\link[=shipAreaFromMass]{shipAreaFromMass()}} with
the provided value of \code{ms}.}

\item{Ly}{Ship impact horizontal extent (m); defaults to 1.23m if not specified,
based on an analysis of the shape of the bow of typical coastal fishing boats
of the Cape Islander variety.}

\item{Lz}{Ship impact vertical extent (m); defaults to 1.23m if not specified,
based on the same analysis as for Ly.}

\item{species}{A string indicating the whale species. For the permitted values,
see \code{\link[=whaleMassFromLength]{whaleMassFromLength()}}.}

\item{lw}{Whale length (m). This is used by \code{\link[=whaleAreaFromLength]{whaleAreaFromLength()}} to
calculate area, which is needed for the water drag calculation done by
\code{\link[=whaleWaterForce]{whaleWaterForce()}}.}

\item{mw}{Whale mass (kg). If this value is not provided, then
it is calculated from whale length, using \code{\link[=whaleMassFromLength]{whaleMassFromLength()}}
with \code{type="wetted"}.}

\item{Sw}{Whale surface area (m^2). If not provided, this is calculated
from whale length using \code{\link[=whaleAreaFromLength]{whaleAreaFromLength()}}.}

\item{l}{Numerical vector of length 4, giving thickness (m) of skin, blubber,
sublayer, and bone. If not provided, this is set to
\code{c(0.025, 0.16, 1.12, 0.1)}.
The skin thickness default of 0.025 m represents the 0.9-1.0 inch value
stated in Section 2.2.3 of Raymond (2007).
The blubber default of 0.16 m is a rounded average of the values inferred
by whale necropsy, reported in Appendix 2 of Daoust et al., 2018.
The sublayer default of 1.12 m may be reasonable at some spots on the whale body.
The bone default of 0.1 m may be reasonable at some spots on the whale body.
The sum of these default values, 1.40 m, is a whale radius that
is consistent with a half-circumference of 4.4 m, reported in Table 2.2
of Raymond (2007).}

\item{a}{Numerical vectors of length 4, giving values to use in the
stress-strain law \code{stress=a*(exp(b*strain)-1)}, where \code{a} is in Pa
and \code{b} is unitless. By construction, \code{a*b} is the local modulus at
low strain (i.e. at low \code{b*strain} values), and that \code{b} is the
efolding scale for nonlinear increase in stress with strain.
This exponential relationship has been mapped out
for whale blubber, using a curve fit to Figure 2.13 of Raymond (2007), and
these values are used for the second layer (blubber); see
the documentation for the \link{raymond2007} dataset, to see
for how that fit was done.
If not provided, \code{a} defaults to
\code{c(17.8e6/0.1, 1.58e5, 1.58e5, 8.54e8/0.1)}
and \code{b} defaults to
\code{c(0.1, 2.54, 2.54, 0.1)}.
The skin defaults are set up to give a linear shape (since \code{b} is small)
with the \code{a*b} product
being 17.8e6 Pa, which is the adult-seal value
given in Table 3 of Grear et al. (2017).
The blubber defaults are from a regression of the stress-strain
relationship shown in Figure 2.13 of Raymond (2007).
The sublayer defaults are set to match those of blubber, lacking
any other information.
The bone default for \code{b} is small, to set up a linear function,
and \code{a*b} is set to equal 8.54e8 Pa,
given in Table 2.3 of Raymond (2007) and Table 4.5 of
Campbell-Malone (2007).}

\item{b}{Numerical vectors of length 4, giving values to use in the
stress-strain law \code{stress=a*(exp(b*strain)-1)}, where \code{a} is in Pa
and \code{b} is unitless. By construction, \code{a*b} is the local modulus at
low strain (i.e. at low \code{b*strain} values), and that \code{b} is the
efolding scale for nonlinear increase in stress with strain.
This exponential relationship has been mapped out
for whale blubber, using a curve fit to Figure 2.13 of Raymond (2007), and
these values are used for the second layer (blubber); see
the documentation for the \link{raymond2007} dataset, to see
for how that fit was done.
If not provided, \code{a} defaults to
\code{c(17.8e6/0.1, 1.58e5, 1.58e5, 8.54e8/0.1)}
and \code{b} defaults to
\code{c(0.1, 2.54, 2.54, 0.1)}.
The skin defaults are set up to give a linear shape (since \code{b} is small)
with the \code{a*b} product
being 17.8e6 Pa, which is the adult-seal value
given in Table 3 of Grear et al. (2017).
The blubber defaults are from a regression of the stress-strain
relationship shown in Figure 2.13 of Raymond (2007).
The sublayer defaults are set to match those of blubber, lacking
any other information.
The bone default for \code{b} is small, to set up a linear function,
and \code{a*b} is set to equal 8.54e8 Pa,
given in Table 2.3 of Raymond (2007) and Table 4.5 of
Campbell-Malone (2007).}

\item{s}{Numerical vector of length 4, giving the ultimate strengths (Pa) of
skin, blubber, sublayer, and bone, respectively. If not provided, the
value is set to \code{1e6 * c(19.600,0.255,0.255,22.900)}
with reasoning as follows.
The skin default of 19.6 MPa
is a rounded value from Table 3 of Grear et al. (2018) for adult seal skin strength at
an orientation of 0 degrees.  The blubber and sublayer values were chosen
as the central point of a logistic fit of whale collision damage
to maximal stress during a default impact simulation.
(For comparison, a strength of
0.437 MPa may be inferred by
multiplying Raymond's (2007) Figure 2.13 elastic modulus of 0.636 MPa
by the ratio 0.97/1.41 determined for adult seal strength/modulus, as reported
in Table 3 of Grear et al. (2018).)
The bone default o 22.9 MPa is from Table 2.3 of Raymond (2007) and
Table 4.5 of Campbell-Malone (2007).}

\item{theta}{Whale skin deformation angle (deg); defaults to 55 degrees,
if not supplied, because that angle produces a good match to Raymond's (2007)
Figure 6.1 for the total force as a function of vessel speed, for large
vessels. (Note that the match works almost as well in the range 50 deg
to 70 deg.)}

\item{Cs}{Drag coefficient for ship (dimensionless),
used by \code{\link[=shipWaterForce]{shipWaterForce()}} to estimate ship drag force. Defaults
to 1e-2, which is 4 times the frictional coefficient of 2.5e-3
inferred from Figure 4 of Manen and van Oossanen (1988), assuming
a Reynolds number of 5e7, computed from speed 5m/s, lengthscale 10m
and viscosity 1e-6 m^2/s. (The factor of 4 is under the assumption
that frictional drag is about a quarter of total drag.)
The drag force is computed with \code{\link[=shipWaterForce]{shipWaterForce()}}.}

\item{Cw}{Drag coefficient for whale (dimensionless),
used by \code{\link[=whaleWaterForce]{whaleWaterForce()}} to estimate whale drag force.
Defaults to 2.5e-3, for Reynolds number 2e7, computed from speed
2 m/s, lengthscale 5m (between radius and length) and
viscosity 1e-6 m^2/s.  The drag force is computed with
\code{\link[=whaleWaterForce]{whaleWaterForce()}}.}

\item{logistic}{a \link{list} containing \code{logStressCenter} and \code{logStressWidth},
which define an empirical logistic fit of an index of whale injury in
observed strikes (ranging from 0 for no injury to 1 for fatal injury),
as a function of the base-10 logarithm of compressive
stress, as well as \code{tau25}, \code{tau50} and \code{tau75}, which are the stresses
in that fit that yield index values of 0.25, 0.50 and 0.75, respectively;
these values set colour boundaries in \code{\link[=plot.strike]{plot.strike()}} plots that have
\code{which="threat"}.}
}
\value{
A named list holding the parameters, with defaults and alternatives reconciled
according to the system described above, along with a function that computes
compression force, which is created by \code{\link[=stressFromStrainFunction]{stressFromStrainFunction()}}.
}
\description{
\code{\link{updateParameters}} is used to alter one or more components of an existing
object of type \code{"parameters"} that was created by \code{\link[=parameters]{parameters()}}. This
can be useful for e.g. sensitivity tests.
}
\details{
The following points must be kept in mind.
\enumerate{
\item Unlike \code{\link[=parameters]{parameters()}}, \code{updateParameters} does not check its arguments
for feasible values.  This can lead to bad results when using
\code{\link[=strike]{strike()}}, which is e.g. expecting four layer thicknesses to
be specified, and also that each thickness is positive.
\item Unlike \code{\link[=parameters]{parameters()}}, \code{updateParameters} does not any ancillary
actions with the supplied values.  For example, \code{\link[=parameters]{parameters()}}
will infer whale mass \code{mw} from whale length \code{lw}, if only
the latter is supplied. The only ancillary actions taken by
\code{updateParameters} is to update an internal variable named
\code{lsum} (as the sum of the components of \code{l}), and also to
store the return value from a call to \code{\link[=stressFromStrainFunction]{stressFromStrainFunction()}}
using the values of \code{l}, \code{a} and \code{b}.
}
}
\author{
Dan Kelley
}
