% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whale_strike}
\alias{whale_strike}
\title{Simulate the collision of a ship and a whale}
\usage{
whale_strike(t, state, parms)
}
\arguments{
\item{t}{time [s].}

\item{state}{A named vector describing the initial state of the model. This
must contain ship position \code{xs} [m],
ship speed \code{vs} [m/s], whale position \code{xw} [m],
and whale speed \code{vw} [m/s].}

\item{parms}{a named vector holding model parameters.
ship mass \code{"ms"} [kg],
ship beam \code{"beam"} [m],
ship draft \code{"draft"} [m],
whale mass \code{"mw"} [kg],
whale skin thickness \code{"delta"} [m],
whale skin elastic modulus \code{"Eskin"} [Pa],
whale skin deformation extension \code{"gamma"} [m],
whale blubber thickness \code{"B"} [m],
and \code{"blubbermodel"} (one of \code{"linear"}, \code{"exponential"}
or \code{"raymond"}).}
}
\value{
An object of class \code{"whalestrike"}, consisting of a
list containing vectors for time (\code{t} [s]), ship position (\code{xs} [m]),
boat speed (\code{vs} [m/s]), whale position (\code{xw} [m]), whale speed (\code{vw} [m/s]),
boat acceleration (\code{dvsdt} [m/s^2]), and whale acceleration (\code{dvwdt} [m/s^2]),
along with a list containing the model parameters (\code{parms}).
}
\description{
Simulate the collision of a ship and a whale
}
\examples{
library(whalestrike)
t <- seq(0, 1, length.out=500)
state <- c(xs=-1.5, vs=5, xw=0, vw=0)
parms <- list(ms=20e3, beam=3, draft=3,
              mw=20e3,
              delta=0.02, Eskin=2e7, gamma=0.2,
              B=0.3, blubbermodel="exponential")
sol <- whale_strike(t, state, parms)
par(mfcol=c(3, 3), mar=c(2, 3, 1, 0.5), mgp=c(2, 0.7, 0), cex=0.7)
plot(sol, which="all")
}
