% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleAreaFromLength}
\alias{whaleAreaFromLength}
\title{Whale projected area, as function of length}
\usage{
whaleAreaFromLength(L, view = "side")
}
\arguments{
\item{L}{length in m}

\item{view}{character string indicating the viewpoint; only
\code{"side"} is permitted at present, because that is what
is needed by \code{\link{whaleWaterForce}}.}
}
\description{
Whale projected area, as function of length
}
