% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleBlubberForce}
\alias{whaleBlubberForce}
\title{Blubber force}
\usage{
whaleBlubberForce(xs, xw, parms)
}
\arguments{
\item{xs}{Ship position [m]}

\item{xw}{Whale position [m]}

\item{parms}{A named list holding model parameters, as created by
\code{\link{parameters}}.}
}
\value{
Compression-resisting force [N]
}
\description{
The stress model depends on \code{parms$blubbermodel}, a character
string with the following possibilities:
\itemize{
\item \code{"Raymond"}: Raymond's (2007) modulus is used, with
stress [Pa] taken to be 636273*strain, where strain is the relative
compression of the blubber.
\item \code{"linear"}: a similar
linear law is used, but with coefficient 580004.6, resulting from a linear
fit to the data in his figure 2.13, constrained to have zero stress for zero
strain.
\item \code{"exponential"}: an exponential model is used, with
stress=-1.712240e5*(1-exp(strain/0.4185693)). (This is a fit to the
data in Raymond's (2007) Figure 2.13.)
}
}
\references{
See \link{whalestrike} for a list of references.
}
