% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleCompressionForce}
\alias{whaleCompressionForce}
\title{Whale compression force}
\usage{
whaleCompressionForce(xs, xw, parms)
}
\arguments{
\item{xs}{Ship position [m]}

\item{xw}{Whale position [m]}

\item{parms}{A named list holding model parameters, created by
\code{\link{parameters}}.}
}
\value{
Compression-resisting force of whale blubber and the layer underneath it [N].
}
\description{
Calculate the reaction force of blubber, as the product of stress
and area. Stress is computed as the product of blubber elastic
modulus \code{parms$Ebeta} times strain, where the latter is computed
from ship-whale separation using \code{1-(xw-xs)/parms$beta} if
\code{xw} is between \code{xs} and \code{xs+parms$beta},
or zero otherwise. Area is computed as the product of
\code{parms$B} and \code{parms$D}.
FIXME: rewrite docs for the new blubber+inner scheme
}
\references{
See \link{whalestrike} for a list of references.
}
