% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleLengthFromMass}
\alias{whaleLengthFromMass}
\title{Compute whale length from mass}
\usage{
whaleLengthFromMass(M, species = "N. Atl. Right Whale",
  model = "fortune2012atlantic")
}
\arguments{
\item{M}{Whale mass (kg).}

\item{species}{A string indicating the whale species. For the permitted values,
see \code{\link[=whaleMassFromLength]{whaleMassFromLength()}}.}

\item{model}{Character string specifying the model, with permitted
values \code{"moore2005"}, \code{"fortune2012atlantic"} and
for \code{\link[=whaleMassFromLength]{whaleMassFromLength()}} for the details of these
formulations.}
}
\value{
Whale length (m).
}
\description{
This works by inverting \code{\link[=whaleMassFromLength]{whaleMassFromLength()}} using \code{\link[=uniroot]{uniroot()}}.
}
\references{
See \code{\link[=whalestrike]{whalestrike()}} for a list of references.
}
\author{
Dan Kelley
}
