% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleMassFromLength}
\alias{whaleMassFromLength}
\title{Whale mass inferred from length}
\usage{
whaleMassFromLength(L, species = "N. Atl. Right Whale",
  model = "fortune2012atlantic")
}
\arguments{
\item{L}{Whale length in m.}

\item{species}{String specifying the species. This must be one of the following:
\code{"N. Atl. Right Whale"} (the default), FIXME: add more.}

\item{model}{Character string specifying the model (see \dQuote{Details}).}
}
\value{
Mass in kg.
}
\description{
Calculate an estimate of the mass of a whale, based on its length, with
formulae from the following sources.
}
\details{
The permitted values for \code{model} are as follows.
\itemize{
\item \code{"moore2005"} yields
\eqn{242.988 * exp(0.4 * length)}{242.988 * exp(0.4 * L)},
which (apart from a unit change on \code{L}) is the regression equation
shown above Figure 1d in Moore et al. (2005) for right whales. A
difficult in the Moore et al. (2005) use of a single nonzero digit
in the multiplier on \code{L} is illustrated in \dQuote{Examples}.

\item \code{"fortune2012atlantic"} yields the formula
\eqn{exp(-10.095 + 2.825*log(100*L))}{exp(-10.095 + 2.825*log(100*L))}
for North Atlantic right whales, according to corrected version of the
erroneous formula given in the caption of Figure 4 in Fortune et al (2012).
(The error, an exchange of slope and intercept, was confirmed by
S. Fortune in an email to D. Kelley dated June 22, 2018.)

\item \code{"fortune2012pacific"} yields the formula
\eqn{exp(-12.286 + 3.158*log(100*L))}{exp(-12.286 + 3.158*log(100*L))}
for North Pacific right whales, according to corrected version of the
erroneous formula given in the caption of Figure 4 in Fortune et al (2012).
(The error, an exchange of slope and intercept, was confirmed by
S. Fortune in an email to D. Kelley dated June 22, 2018.)
}
}
\examples{
library(whalestrike)
L <- seq(5, 15, length.out=100)
kpt <- 1000 # kg per tonne
plot(L, whaleMassFromLength(L, model="moore2005")/kpt, type='l',
     xlab="Right-whale Length [m]", ylab="Mass [tonne]")
# Demonstrate sensitivity involved in the single-digit parameter in Moore's formula
lines(L, 242.988 * exp(0.35 * L)/kpt, lty='dotted')
lines(L, 242.988 * exp(0.45 * L)/kpt, lty='dashed')
lines(L, whaleMassFromLength(L, model="fortune2012atlantic")/kpt, col=2)
lines(L, whaleMassFromLength(L, model="fortune2012pacific")/kpt, col=3)
legend("topleft", lwd=1, col=1:3,
       legend=c("moore2005", "fortune2012atlantic", "fortune2012pacific"))

}
\references{
See \link{whalestrike} for a list of references.
}
