% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forces.R
\name{whaleSkinForce}
\alias{whaleSkinForce}
\title{Skin force}
\usage{
whaleSkinForce(xs, xw, parms)
}
\arguments{
\item{xs}{Ship position (m).}

\item{xw}{Whale position (m).}

\item{parms}{A named list holding model parameters, created by
\code{\link[=parameters]{parameters()}}.}
}
\value{
A list containing \code{force}, the normal force (N), along with
\code{sigmay} and \code{sigmaz}, which are stresses (Pa) in the y (beam)
and z (draft) directions.
}
\description{
The ship-whale separation is used to calculate the deformation of the skin. The
parameters of the calculation are \code{parms$Ly} (impact area width, m),
\code{parms$Lz} (impact area height, in m), \code{parms$Ealpha} (skin elastic modulus in Pa),
\code{parms$alpha} (skin thickness in m), and \code{parms$theta} (skin bevel angle
degrees, measured from a vector normal to undisturbed skin).
}
\references{
See \code{\link[=whalestrike]{whalestrike()}} for a list of references.
}
\seealso{
Other functions relating to whale characteristics: 
\code{\link{stressFromStrainFunction}()},
\code{\link{whaleCompressionForce}()},
\code{\link{whaleLengthFromMass}()},
\code{\link{whaleMassFromLength}()},
\code{\link{whaleWaterForce}()}

Other functions relating to forces: 
\code{\link{shipWaterForce}()},
\code{\link{stressFromStrainFunction}()},
\code{\link{whaleCompressionForce}()},
\code{\link{whaleWaterForce}()}
}
\author{
Dan Kelley
}
\concept{functions relating to forces}
\concept{functions relating to whale characteristics}
