% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\docType{package}
\name{whalestrike}
\alias{whalestrike}
\alias{whalestrike-package}
\title{whalestrike: A Package to Simulate Ship-Whale Collisions}
\description{
This package solves Newton's second law for a simple model of
a ship colliding with a whale. This is a stripped-down model
that does not attempt to simulate the biomechanical interactions
that are possible in the finite-element treatment of Raymond (2007).
Rather, the goal is to establish a convenient framework for rapid
computation of responses to a variety of conditions. Indeed, the model runs
so quickly that it easily keeps up with mouse movements in an R shiny
app, in which the user can test the effects of e.g. changing contact
area, ship speed, etc., to build intuition on scenarios ranging
from collision with a slow-moving fishing boat to collision
with a thin dagger-board of a rapidly-moving racing sailboat.
The simple formulation also makes it easy to modify the system
by altering various dynamical and biomechanical parameters,
and by adding new forces or criteria for whale damage.
}
\details{
See the documentation for \code{\link{strike}} for a typical
example of using the main functions of this package.
}
\references{
\itemize{

\item
Fortune, Sarah M. E., Andrew W. Trites, Wayne L. Perryman, Michael J. Moore,
Heather M. Pettis, and Morgan S. Lynn. “Growth and Rapid Early Development of
North Atlantic Right Whales (Eubalaena Glacialis).” Journal of Mammalogy 93,
no. 5 (2012): 1342–54. https://doi.org/10.1644/11-MAMM-A-297.1.

\item
Grear, Molly E., Michael R. Motley, Stephanie B. Crofts, Amanda E. Witt, Adam
P. Summers, and Petra Ditsche. “Mechanical Properties of Harbor Seal Skin and
Blubber − a Test of Anisotropy.” Zoology 126 (2018): 137–44.
https://doi.org/10.1016/j.zool.2017.11.002.

\item
Miller, Carolyn A., Desray Reeb, Peter B. Best, Amy R. Knowlton, Moira W.
Brown, and Michael J. Moore. “Blubber Thickness in Right Whales Eubalaena
Glacialis and Eubalaena Australis Related with Reproduction, Life History
Status and Prey Abundance.” Marine Ecology Progress Series 438 (2011): 267–83.

\item
Moore, M.J., A.R. Knowlton, S.D. Kraus, W.A. McLellan, and R.K. Bonde.
“Morphometry, Gross Morphology and Available Histopathology in North Atlantic
Right Whale (Eubalaena Glacialis) Mortalities (1970 to 2002).” Journal of
Cetacean Research and Management 6, no. 3 (2005): 199–214.

\item
Ng, Laurel J., Vladislav Volman, Melissa M. Gibbons, Pi Phohomsiri, Jianxia
Cui, Darrell J. Swenson, and James H. Stuhmiller. “A Mechanistic End-to-End
Concussion Model That Translates Head Kinematics to Neurologic Injury.”
Frontiers in Neurology 8, no. JUN (2017): 1–18.
\url{https://doi.org/10.3389/fneur.2017.00269}

\item
Raymond, J. J. “Development of a Numerical Model to Predict Impact Forces on a
North Atlantic Right Whale during Collision with a Vessel.” University of New
Hampshire, 2007. https://scholars.unh.edu/thesis/309.
}
}
