% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Run a GUI app for interactive simulations}
\usage{
app(mode = "simple", options = list(height = 800))
}
\arguments{
\item{mode}{Character string specifying the style to use.  Only
the value \code{"simple"} is permitted at present. This yields a
3-panel plot, constructed by \code{\link{plot.strike}},
called with default arguments.}

\item{options}{List containing options that are provided
to \code{\link[shiny]{shinyApp}}, which creates the GUI app.}
}
\description{
Run a GUI app for interactive simulations
}
