% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app2.R
\name{app2}
\alias{app2}
\title{GUI application for whale simulation (version 2)}
\usage{
app2(debug = FALSE)
}
\arguments{
\item{debug}{logical value indicating whether to print output to
the R console as the computation is done.}
}
\description{
This is similar to \code{\link[=app]{app()}}, except that it relies on the \code{bslib} package
to provide a cleaner interface, in which sub-windows for controllers
can be opened and closed by the user. Since \code{bslib} is still experimental,
\code{\link[=app2]{app2()}} may start failing in the future, and users are asked to
report an issue on the package github page, so the author
can make adjustments.
}
\details{
Compared with \code{\link[=app]{app()}}, the present function lacks the ability to save settings
and reload them later. This is mainly because it only works with locally-run
operations, not from server-run operations.  The latter would require extra
coding to set up user's storage space, to prevent against web attacks, etc.,
which is beyond the present purpose.  However, there is an addition with
\code{\link[=app2]{app2()}} that might prove more useful: a button to display the code required
to reproduce the simulated state.  This may be of help to the those seeking
to explore the simulator in programmatically, for precision and reproducibility.

Sliders, buttons, and choosers are grouped into panes that appear on
the left of the view. When \code{\link[=app2]{app2()}} first opens, all of these panes
are closed. To get acquainted with the app, try adjusting the controllers
that \emph{are} visible on the initial view.  Then, open the "ship" pane and
increase the ship mass.  Do you find that the results make qualitative
sense?  Continue this process, exploring all the panes. It is hoped
that a half hour of such exploration will let users start to
investigate practical applications.  For more about how the simulations
are carried out, as well as comments on some applications that
may be of interest, please consult Kelley et al. (2020).
}
\references{
Kelley, Dan E., James P. Vlasic, and Sean W. Brillant. "Assessing the Lethality
of Ship Strikes on Whales Using Simple Biophysical Models." Marine Mammal
Science, October 12, 2020, mms.12745. \doi{10.1111/mms.12745}.
}
\seealso{
Other interactive apps: 
\code{\link{app}()}
}
\author{
Dan Kelley
}
\concept{interactive apps}
