% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{contactArea}
\alias{contactArea}
\title{Contact area}
\usage{
contactArea(xs, xw, parms)
}
\arguments{
\item{xs}{Ship position [m]}

\item{xw}{Whale position [m]}

\item{parms}{A named list holding model parameters, created by
\code{\link{parameters}}.}
}
\value{
Contact area [m^2].
}
\description{
Area of contact between vessel and whale skin
}
\section{DEVELOPMENT NOTE}{
 think about formulation (re theta).
}

