% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{derivative}
\alias{derivative}
\title{Calculate derivative using first difference}
\usage{
derivative(var, t)
}
\arguments{
\item{var}{variable.}

\item{t}{time in seconds.}
}
\value{
Derivative estimated by using \code{\link{diff}} on both \code{x}
and \code{t}. In order to return a value of the same length as \code{x} and
\code{t}, the last value is repeated.
}
\description{
Calculate derivative using first difference
}
\author{
Dan Kelley
}
