% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{dynamics}
\alias{dynamics}
\title{Dynamical law}
\usage{
dynamics(t, y, parms)
}
\arguments{
\item{t}{time (s).}

\item{y}{model state, a vector containing ship position \code{xs} (m),
ship speed \code{vs} (m/s), whale position \code{xw} (m),
and whale speed \code{vw} (m/s).}

\item{parms}{A named list holding model parameters, created by
\code{\link{parameters}}.}
}
\description{
This function handles Newton's second law, which is the dynamical
law that relates the accelerations of whale and ship to the forces
upon each.  It is used by \code{\link[=strike]{strike()}}, as the latter integrates
the acceleration equations to step forward in time through
the simulation of a whale-strike event. Thus, \code{\link[=dynamics]{dynamics()}}
is a core function of this package.  The code is very simple,
because the forces are determined by other functions, as
described in the \dQuote{Details} section.
}
\details{
Given a present state (defined by the positions and
velocities of ship and whale) at the present time,
apply Newton's second law to find the time derivatives
of that state.  Forces are determined with
\code{\link[=whaleCompressionForce]{whaleCompressionForce()}},
\code{\link[=whaleExtensionForce]{whaleExtensionForce()}},
\code{\link[=shipWaterForce]{shipWaterForce()}},
\code{\link[=whaleWaterForce]{whaleWaterForce()}}, while engine force
(assumed constant over the course of a collision) is
computed from initial \code{\link[=shipWaterForce]{shipWaterForce()}}.  Whale and
ship masses are set by \code{\link[=parameters]{parameters()}}, which also sets up
areas, drag coefficients, etc.
}
\references{
See \code{\link[=whalestrike]{whalestrike()}} for a list of references.
}
\author{
Dan Kelley
}
