% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{dynamics}
\alias{dynamics}
\title{Dynamical law}
\usage{
dynamics(t, y, parms)
}
\arguments{
\item{t}{time [s].}

\item{y}{model state, a vector containing ship position \code{xs} [m],
ship speed \code{vs} [m/s], whale position \code{xw} [m],
and whale speed \code{vw} [m/s].}

\item{parms}{model parameters.}
}
\description{
Dynamical law
}
