% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{massFromLength}
\alias{massFromLength}
\title{Whale mass inferred from length}
\usage{
massFromLength(L, model = "fortune2012atlantic")
}
\arguments{
\item{L}{Whale length in m.}

\item{model}{Character string specifying the model (see \dQuote{Details}).}
}
\value{
Mass in kg.
}
\description{
Calculate an estimate of the mass of a whale, based on its length.
}
\details{
The permitted values for \code{model} are as follows.
\itemize{
\item \code{"moore2005"} yields
\eqn{242.988 * exp(0.4 * length)}{242.988 * exp(0.4 * L)},
which (apart from a unit change on \code{L}) is the regression equation
shown above Figure 1d in Moore et al. (2005) for right whales. A
difficult in the Moore et al. (2005) use of a single nonzero digit
in the multiplier on \code{L} is illustrated in \dQuote{Examples}.

\item \code{"fortune2012atlantic"} yields the formula
\eqn{exp(-10.095 + 2.825*log(100*L))}{exp(-10.095 + 2.825*log(100*L))}
for North Atlantic right whales, according to corrected version of the
erroneous formula given in the caption of Figure 4 in Fortune et al (2012).
(The error, an exchange of slope and intercept, was confirmed by
S. Fortune in an email to D. Kelley dated June 22, 2018.)

\item \code{"fortune2012pacific"} yields the formula
\eqn{exp(-12.286 + 3.158*log(100*L))}{exp(-12.286 + 3.158*log(100*L))}
for North Pacific right whales, according to corrected version of the
erroneous formula given in the caption of Figure 4 in Fortune et al (2012).
(The error, an exchange of slope and intercept, was confirmed by
S. Fortune in an email to D. Kelley dated June 22, 2018.)
}
}
\examples{
library(whalestrike)
L <- seq(5, 15, length.out=100)
kpt <- 1000 # kg per tonne
plot(L, massFromLength(L, model="moore2005")/kpt, type='l',
     xlab="Right-whale Length [m]", ylab="Mass [tonne]")
# Demonstrate sensitivity involved in the single-digit parameter in Moore's formula
lines(L, 242.988 * exp(0.35 * L)/kpt, lty='dotted')
lines(L, 242.988 * exp(0.45 * L)/kpt, lty='dashed')
lines(L, massFromLength(L, model="fortune2012atlantic")/kpt, col=2)
lines(L, massFromLength(L, model="fortune2012pacific")/kpt, col=3)
legend("topleft", lwd=1, col=1:3,
       legend=c("moore2005", "fortune2012atlantic", "fortune2012pacific"))
}
\references{
Moore, M.J., A.R. Knowlton, S.D. Kraus, W.A. McLellan, and R.K. Bonde.
“Morphometry, Gross Morphology and Available Histopathology in North Atlantic
Right Whale (Eubalaena Glacialis) Mortalities (1970 to 2002).” Journal of
Cetacean Research and Management 6, no. 3 (2005): 199–214.

Fortune, Sarah M. E., Andrew W. Trites, Wayne L. Perryman, Michael J. Moore,
Heather M. Pettis, and Morgan S. Lynn. “Growth and Rapid Early Development of
North Atlantic Right Whales (Eubalaena Glacialis).” Journal of Mammalogy 93,
no. 5 (2012): 1342–54. https://doi.org/10.1644/11-MAMM-A-297.1.
}
