% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lethality_index.R
\name{maximumLethalityIndex}
\alias{maximumLethalityIndex}
\title{Compute maximum Lethality Index for a strike simulation}
\usage{
maximumLethalityIndex(strike)
}
\arguments{
\item{strike}{the value returned by a call to strike.}
}
\description{
Compute maximum Lethality Index for a strike simulation
}
\examples{
library(whalestrike)
t <- seq(0, 0.7, length.out = 200)
state <- list(xs = -2, vs = knot2mps(10), xw = 0, vw = 0)
parms <- parameters()
s <- strike(t, state, parms)
max(lethalityIndexFromStress(s[["WCF"]][["stress"]]))
maximumLethalityIndex(s)

}
\seealso{
Other functions dealing with Whale Lethality index: 
\code{\link{lethalityIndexFromStress}()},
\code{\link{stressFromLethalityIndex}()}
}
\author{
Dan Kelley, wrapping code provided by Alexandra Mayette
}
\concept{functions dealing with Whale Lethality index}
