% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{parameters}
\alias{parameters}
\title{Control parameters for whale strike simulation}
\usage{
parameters(ms, as = 20, B = 3, D = 1.5, mw, lw, delta = 0.01, Es,
  theta = 45, beta = 0.3, blubbermodel = "nonlinear", CDs = 1,
  CDw = 1)
}
\arguments{
\item{ms}{Ship mass [kg].}

\item{as}{Ship area [m^2]. This, together with \code{CDs}, is used by
used by \code{\link{shipWaterForce}} to estimate ship drag force.}

\item{B}{Ship impact horizontal extent [m]; defaults to 2m if not specified.}

\item{D}{Ship impact vertical extent [m]; defaults to 1.5m if not specified.}

\item{mw}{Whale mass [kg]. If not supplied, \code{\link{whaleMassFromLength}} is used
to calculate this, given \code{lw}, but if neither \code{mw} nor \code{ml} is provided,
an error is reported.}

\item{lw}{Whale length [m]. If not supplied, \code{\link{whaleLengthFromMass}} is used
to calculate this, given \code{lm}, but if neither \code{mw} nor \code{ml} is provided,
an error is reported. The length is used by \code{\link{whaleAreaFromLength}} to
calculate area, which is needed for the water drag calculation done by
\code{\link{whaleWaterForce}}.}

\item{delta}{Whale skin thickness [m]. Defaults to 0.01 m, if not supplied.}

\item{Es}{Whale skin elastic modulus [Pa]. If not provided, defaults to 20e6 Pa,
a rounded value of the 19.56+-4.03MPa estimate for seal skin, provided in Table 3
of Grear et al. (2017).}

\item{theta}{Whale skin deformation angle [deg]; defaults to 45deg if not supplied.}

\item{beta}{Whale blubber thickness [m]; defaults to 0.3m if not supplied.}

\item{blubbermodel}{Blubber model (see \code{\link{whaleBlubberForce}}); defaults to \code{"nonlinear"}
if not supplied.}

\item{CDs}{Drag coefficient for ship [dimensionless], used by \code{\link{shipWaterForce}} to estimate ship drag force.}

\item{CDw}{Drag coefficient for whale [dimensionless], used by \code{\link{whaleWaterForce}} to estimate whale drag force.}
}
\value{
A named list holding the parameters, with defaults and alternatives reconciled
according to the system described above.
}
\description{
Assembles control parameters into a list suitable for passing to \code{\link{strike}}
and the functions that it calls.
}
\references{
See \link{whalestrike} for a list of references.
}
