% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{parameters}
\alias{parameters}
\title{Control parameters for whale strike simulation}
\usage{
parameters(ms, Ss, impactWidth = 3, impactHeight = 1.5, lw, mw, Sw,
  alpha = 0.0256, Ealpha = 17800000, UTSalpha = 19560000, theta = 45,
  beta = 0.3, Ebeta = 6e+05, UTSbeta = (0.8/1.2) * 6e+05, gamma = 0.5,
  Egamma = 4e+05, UTSgamma = (0.8/1.2) * 4e+05, Cs = 0.01, Cw = 0.0025,
  file)
}
\arguments{
\item{ms}{Ship mass [kg].}

\item{Ss}{Ship wetted area [m^2]. This, together with \code{Cs}, is used by
used by \code{\link{shipWaterForce}} to estimate ship drag force. If \code{Ss}
is not given, then an esimate is made by calling \code{\link{shipAreaFromMass}} with
the provided value of \code{ms}.}

\item{impactWidth}{Ship impact horizontal extent [m]; defaults to 2m if not specified.}

\item{impactHeight}{Ship impact vertical extent [m]; defaults to 1.5m if not specified.}

\item{lw}{Whale length [m]. If not supplied, \code{\link{whaleLengthFromMass}} is used
to calculate this, given \code{lm}, but if neither \code{mw} nor \code{ml} is provided,
an error is reported. The length is used by \code{\link{whaleAreaFromLength}} to
calculate area, which is needed for the water drag calculation done by
\code{\link{whaleWaterForce}}.}

\item{mw}{Whale mass [kg]. If not provided, this is calculated from whale
length, using \code{\link{whaleMassFromLength}} with \code{type="wetted"}.}

\item{Sw}{Whale surface area [m^2]. If not provided, this is calculated
from whale length using \code{\link{whaleAreaFromLength}}.}

\item{alpha}{Whale skin thickness [m], with a of 0.0256 m.}

\item{Ealpha}{Whale skin elastic modulus [Pa], with a default of 17.80e6 Pa,
a value for adult seals, given in Table 3 of Grear et al. (2017).}

\item{UTSalpha}{Ultimate tensile strength of skin [Pa], with a default of 19.56e6 Pa,
a value reported by Grear et al. (2017) for adult seals.}

\item{theta}{Whale skin deformation angle [deg]; defaults to 45deg if not supplied.}

\item{beta}{Whale blubber thickness [m]; defaults to 0.3m if not supplied.}

\item{Ebeta}{Elastic modulus of blubber [Pa]; defaults to 0.6 MPa (the value
suggested Raymond (2007 fig 37), rounded to 1 digit), if not supplied.}

\item{UTSbeta}{Numerical value indicating the ultimate tensile strength
of the blubber; if not \code{NA}, then this is used to indicate
problems in plots made if \code{which} is \code{"compression stress"}.
The default is the product of the default whale blubber modulus (see
\code{Ebeta}, above) and the strength/modulus ratio for seal
blubber, given by (Grear et al. 2018 #' page 144).}

\item{gamma}{Thickness of interior region [m].
The default, 0.5m, may be in an appropriate range for soft tissue;
perhaps 0.05m would be more reasonable for bone.}

\item{Egamma}{Elastic modulus of interior region [Pa].
The default, 0.4MPa, may be in an appropriate range for soft tissue;
perhaps 854MPa would be more reasonable for bone.}

\item{UTSgamma}{Numerical value indicating the ultimate tensile strength
of the sublayer.
The default, (0.8/1.2)*0.4MPa, may be in an appropriate range for soft tissue;
perhaps 22.9MPa would be more reasonable for bone.}

\item{Cs}{Drag coefficient for ship [dimensionless],
used by \code{\link{shipWaterForce}} to estimate ship drag force. Defaults
to 1e-2, which is 4 times the frictional coefficient of 2.5e-3
inferred from Figure 4 of Manen and van Oossanen (1988), assuming
a Reynolds number of 5e7, computed from speed 5m/s, lengthscale 10m
and viscosity 1e-6 m^2/s. (The factor of 4 is under the assumption
that frictional drag is about a quarter of total drag.)
The drag force is computed with \code{\link{shipWaterForce}}.}

\item{Cw}{Drag coefficient for whale [dimensionless],
used by \code{\link{whaleWaterForce}} to estimate whale drag force.
Defaults to 2.5e-3, for Reynolds number 2e7, computed from speed
2 m/s, lengthscale 5m (between radius and length) and
viscosity 1e-6 m^2/s.  The drag force is computed with
\code{\link{whaleWaterForce}}.}

\item{file}{Optional name a comma-separated file that holds all of the
previous values, except \code{Cs} and \code{Cw}. If provided,
then other parameters (except \code{Cs} and \code{Cw}) are
ignored, because values are sought from the file. The purpose of
this is in shiny apps that want to save a simulation framework.
The file should be saved \code{\link{write.csv}} with
\code{row.names=FALSE}.}
}
\value{
A named list holding the parameters, with defaults and alternatives reconciled
according to the system described above.
}
\description{
Assembles control parameters into a list suitable for passing to \code{\link{strike}}
and the functions that it calls. If \code{file} is provided, then all the other
arguments are read from that source.
}
\references{
See \link{whalestrike} for a list of references.
}
