% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{parameters}
\alias{parameters}
\title{Control parameters for whale strike simulation}
\usage{
parameters(ms = 20000, Ss, Ly = 0.5, Lz = 1, lw = 13, mw, Sw, l, a, b,
  s, theta = 45, Cs = 0.01, Cw = 0.0025, file)
}
\arguments{
\item{ms}{Ship mass [kg].}

\item{Ss}{Ship wetted area [m^2]. This, together with \code{Cs}, is used by
used by \code{\link{shipWaterForce}} to estimate ship drag force. If \code{Ss}
is not given, then an esimate is made by calling \code{\link{shipAreaFromMass}} with
the provided value of \code{ms}.}

\item{Ly}{Ship impact horizontal extent [m]; defaults to 0.5m if not specified.}

\item{Lz}{Ship impact vertical extent [m]; defaults to 1.5m if not specified.}

\item{lw}{Whale length [m]. If not supplied, \code{\link{whaleLengthFromMass}} is used
to calculate this, given \code{lm}, but if neither \code{mw} nor \code{ml} is provided,
an error is reported. The length is used by \code{\link{whaleAreaFromLength}} to
calculate area, which is needed for the water drag calculation done by
\code{\link{whaleWaterForce}}.}

\item{mw}{Whale mass [kg]. If not provided, this is calculated from whale
length, using \code{\link{whaleMassFromLength}} with \code{type="wetted"}.}

\item{Sw}{Whale surface area [m^2]. If not provided, this is calculated
from whale length using \code{\link{whaleAreaFromLength}}.}

\item{l}{Numerical vector of length 4, giving thickness [m] of skin, blubber,
sublayer, and bone. If not provided, this is set to
\code{c(0.025, 0.16, 0.5, 0.2)}.
The skin thicknes default of 0.025 m represents the 0.9-1.0 inch value
stated in Section 2.2.3 of Raymond (2007).
The blubber default of 0.16 m is a rounded average of the values inferred
by whale necropsy, reported in Appendix 2 of Daoust et al., 2018.
The sublayer default of 0.5 m may be reasonable at some spots on the whale body.
The bone default of 0.1 m may be reasonable at some spots on the whale body.}

\item{a, b}{Numerical vectors of length 4, giving values to use in the
stress-strain law \code{stress=a*(exp(b*strain)-1)}. \code{a} is in Pa
and \code{b} is unitless.  Note that \code{a*b} is the local modulus at
low strain, and that \code{b} is the efolding scale for nonlinear increase
in stress with strain. This exponential relationship has been mapped out
for whale blubber, using a curve fit to Figure 2.13 of Raymond (2007), and
these values are used for the second layer (blubber); see
\link{raymond2007} for how the fit was done.
If not provided, \code{a} defaults to
\code{c(17.80e6/0.1, 1.64e5, 1.64e5, 854.2e6/0.1)}
and \code{b} defaults to
\code{c(0.1, 2.47, 2.47, 0.1)}.
The skin defaults are set up to give a linear shape (since \code{b} is small)
with the \code{a*b} product
being 17.8e6 Pa, which is the adult-seal value
given in Table 3 of Grear et al. (2017).
The blubber defaults are from a regression of the stress-strain
relationship shown in Figure 2.13 of Raymond (2007).
The sublayer defaults are set to match those of blubber, lacking
any other information.
The bone default for \code{b} is small, to set up a linear function,
and \code{a*b} is set to equal 8.54e8 Pa,
given in Table 2.3 of Raymond (2007).}

\item{s}{Numerical vector of length 4, giving the ultimate strengths [Pa] of
skin, blubber, sublayer, and bone, respectively. If not provided, the
value is set to
\code{c(19.56e6, 4.37e5, 4.37e5, 22.9e6)},
with reasoning as follows.
The skin default of 1.96e7 Pa
is a rounded value from Table 3 of Grear et al. (2018) for adult seal skin strength at
an orientation of 0 degrees.  The blubber value of
4.37e5 Pa is inferred by
multiplying Raymond's (2007) Figure 2.13 elastic modulus of 6.36e5 Pa
by the ratio 0.97/1.41 determined for adult seal strength/modulus, as reported
in Table 3 of Grear et al. (2018).
The sublayer value is taken to be identical to the blubber value, lacking
more specific information.
The bone default o 2.29e7 Pa is from Table 2.3 of Raymond (2007).}

\item{theta}{Whale skin deformation angle [deg]; defaults to 45deg if not supplied.}

\item{Cs}{Drag coefficient for ship [dimensionless],
used by \code{\link{shipWaterForce}} to estimate ship drag force. Defaults
to 1e-2, which is 4 times the frictional coefficient of 2.5e-3
inferred from Figure 4 of Manen and van Oossanen (1988), assuming
a Reynolds number of 5e7, computed from speed 5m/s, lengthscale 10m
and viscosity 1e-6 m^2/s. (The factor of 4 is under the assumption
that frictional drag is about a quarter of total drag.)
The drag force is computed with \code{\link{shipWaterForce}}.}

\item{Cw}{Drag coefficient for whale [dimensionless],
used by \code{\link{whaleWaterForce}} to estimate whale drag force.
Defaults to 2.5e-3, for Reynolds number 2e7, computed from speed
2 m/s, lengthscale 5m (between radius and length) and
viscosity 1e-6 m^2/s.  The drag force is computed with
\code{\link{whaleWaterForce}}.}

\item{file}{Optional name a comma-separated file that holds all of the
previous values, except \code{Cs} and \code{Cw}. If provided,
then other parameters (except \code{Cs} and \code{Cw}) are
ignored, because values are sought from the file. The purpose of
this is in shiny apps that want to save a simulation framework.
The file should be saved \code{\link{write.csv}} with
\code{row.names=FALSE}.}
}
\value{
A named list holding the parameters, with defaults and alternatives reconciled
according to the system described above.
}
\description{
Assembles control parameters into a list suitable for passing to \code{\link{strike}}
and the functions that it calls. If \code{file} is provided, then all the other
arguments are read from that source.
Below are some sources cited in the discussion of the function arguments.

Daoust, Pierre-Yves, Émilie L. Couture, Tonya Wimmer, and Laura Bourque.
“Incident Report. North Atlantic Right Whale Mortality Event in the Gulf of St.
Lawrence, 2017.” Canadian Wildlife Health Cooperative, Marine Animal Response
Society, and Fisheries and Oceans Canada, 2018.
\url{http://publications.gc.ca/site/eng/9.850838/publication.html}.

Grear, Molly E., Michael R. Motley, Stephanie B. Crofts, Amanda E. Witt, Adam
P. Summers, and Petra Ditsche. “Mechanical Properties of Harbor Seal Skin and
Blubber − a Test of Anisotropy.” Zoology 126 (2018): 137–44.
\url{https://doi.org/10.1016/j.zool.2017.11.002}.

Raymond, J. J. “Development of a Numerical Model to Predict Impact Forces on a
North Atlantic Right Whale during Collision with a Vessel.” University of New
Hampshire, 2007. \url{https://scholars.unh.edu/thesis/309}.
}
\examples{
parms <- parameters(ms=20e3, lw=13)
epsilon <- seq(0, 0.3, length.out=100)
strain <- parms$stressFromStrain(epsilon)
plot(epsilon, strain, type="l")
}
