% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{plot.strike}
\alias{plot.strike}
\title{Plot a strike object}
\usage{
\method{plot}{strike}(x, which = "all", center = FALSE,
  indicateEvents = FALSE, debug = 0, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{strike}}

\item{which}{A character vector that indicates what to plot.
This choices for its entries are:
\code{"location"} for a time-series plot of boat location \code{xw} in black,
whale location \code{x} in red, and skin location in dashed red,
\code{"whale acceleration"} for a time-series plot of whale acceleration,
\code{"water force"} for a time-series plot of the water-drag force on the whale,
\code{"blubber thickness"} for a time-series plot of blubber thickness,
\code{"blubber force"} for a time-series plot of the normal force resulting from blubber compression,
\code{"blubber stress"} for a time-series plot of the normal stress on the blubber,
\code{"skin force"} for a time-series plot of the normal force resulting from skin tension,
\code{"skin tension"} for a time-series plot of the skin tension, in the along-skin direction,
or \code{"all"} for all of the above.}

\item{center}{Logical, indicating whether to center time-series plots.
on the time when the vessel and whale and in closest proximity.}

\item{indicateEvents}{Logical, indicating whether to draw lines for some events,
such as the moment of closest approach.}

\item{debug}{Integer indicating debugging level, 0 for quiet operation and higher values
for more verbose monitoring of progress through the function.}

\item{...}{Other arguments (ignored).}
}
\description{
See \code{\link{strike}} for examples.
}
