% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{plot.strike}
\alias{plot.strike}
\title{Plot a strike object}
\usage{
\method{plot}{strike}(x, which = "default", drawEvents = TRUE,
  colwcenter = "Slate Gray", colwinterface = "black", colwskin = "black",
  cols = "black", colThreat, lwd = 1.4, D = 3, debug = 0, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{strike}}

\item{which}{A character vector that indicates what to plot.
This choices for its entries are listed below, in no particular order.
\itemize{

\item \code{"location"} for a time-series plot of boat location \code{xw} in
dashed black, whale centerline \code{xs} in solid gray,
blubber-interior interface in red, and skin in blue. The maximum
acceleration of ship and whale (in "g" units) are indicated in notes
placed near the horizontal axes.

\item \code{"section"} to plot skin thickness, blubber thickness and sublayer thickness
in one panel, creating a cross-section diagram.


\item \code{"threat"} a stacked plot showing time-series traces of an
ad-hoc measure of the possible threat to skin, blubber and sublayer.
The threat level is computed as the ratio
of stress to ultimate strength, e.g. for blubber, it is
\code{x$WCF$stress/x$parms$s[2]}. Colour-coding indicates levels
from 0 to 1/2, from 1/2 to 1, and above 1. An axis categorizes
threat level 0 as "L", 0.5 as "M", and 1 as "H".

\item \code{"whale acceleration"} for a time-series plot of whale acceleration.

\item \code{"blubber thickness"} for a time-series plot of blubber thickness.

\item \code{"sublayer thickness"} for a time-series plot of the thickness
of the layer interior to the blubber.

\item \code{"reactive forces"} for a time-series showing the reactive
forces associated with skin stretching (solid) and the compression of the
blubber and sublayer components (dashed).

\item \code{"compression stress"} for a time-series plot of the compression stress on the blubber
and the layer to its interior. (These stresses are equal, under an equilibrium assumption.)

\item \code{"skin stress"} for a time-series of skin stress in the along-skin y and z directions.

\item \code{"values"} for a listing of \code{param} values.

\item \code{"all"} for all of the above.

\item \code{"default"} for a three-element plot showing \code{"location"},
\code{"section"}, and \code{"threat"}.
}}

\item{drawEvents}{Logical, indicating whether to draw lines for some events,
such as the moment of closest approach.}

\item{colwcenter}{Colour used to indicate the whale centre.}

\item{colwinterface}{Colour used to indicate the interface
between whale blubber and sublayer.}

\item{colwskin}{Colour used to indicate the whale skin.}

\item{cols}{As \code{colw}, but the colour to be used for the ship bow location,
which is drawn with a dashed line.}

\item{colThreat}{Three-element colour specification used in \code{"threat"}.
The first colour is used for threat levels between 0 and 0.5, the second
from 0.5 to 1, and the third above 1. If not provided, the colours
will be light-gray, gray, and black, each with an alpha setting that
lightens the colours and makes them semi-transparent.}

\item{lwd}{Line width used in plots for time intervals in which damage
criteria are not exceeded.}

\item{D}{Factor by which to thicken lines during times during which damage
criteria are exceeded.}

\item{debug}{Integer indicating debugging level, 0 for quiet operation and higher values
for more verbose monitoring of progress through the function.}

\item{...}{Ignored.}
}
\description{
Creates displays of various results of a simulation performed
with \code{\link{strike}}.
}
\examples{
## 1. default 3-panel plot
t <- seq(0, 0.7, length.out=200)
state <- c(xs=-2, vs=10*0.5144, xw=0, vw=0) # 10 knot ship
parms <- parameters(ms=20e3, lw=13)
sol <- strike(t, state, parms)
par(mar=c(3,3,1,1), mgp=c(2,0.7,0), mfrow=c(1, 3))
plot(sol)
## 2. all 12 plot types
par(mar=c(3,3,1,1) ,mgp=c(2,0.7,0), mfrow=c(4,3))
plot(sol, "all")
}
\references{
See \link{whalestrike} for a list of references.
}
