% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{plot.strike}
\alias{plot.strike}
\title{Plot a strike object}
\usage{
\method{plot}{strike}(x, which = "default", drawEvents = TRUE,
  colwcenter = "black", colwinterface = "black", colwskin = "black",
  cols = "black", colThreat = c("lightgray", "black"), lwd = 1, D = 3,
  debug = 0, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{strike}}

\item{which}{A character vector that indicates what to plot.
This choices for its entries are listed below, in no particular order.
\itemize{

\item \code{"location"} for a time-series plot of boat location \code{xw} in
dashed black, whale centerline \code{xs} in solid gray,
blubber-interior interface in red, and skin in blue. The maximum
acceleration of ship and whale (in "g" units) are indicated in notes
placed near the horizontal axes. Those acceleration indications report
just a single value for each of ship and whale, but if the blubber
and sublayer have been squeezed to their limits, yielding a short and
intense force spike as the bone compresses, then the summaries will
also report on the spike duration and intensity. The spike is computed
based on using \code{\link{runmed}} on the acceleration data, with a
\code{k} value that is set to correspond to 5 ms, or to k=11, whichever
is larger.

\item \code{"section"} to plot skin thickness, blubber thickness and sublayer thickness
in one panel, creating a cross-section diagram.


\item \code{"threat"} a stacked plot showing time-series traces of an
ad-hoc measure of the possible threat to skin, blubber and sublayer.
The threat level is computed as the ratio
of stress to ultimate strength, e.g. for blubber, it is
\code{x$WCF$stress/x$parms$s[2]}. The vertical same scale is used in
each of the subpanels that make up the stack. Any values exceeding
10 are clipped to 10, and in such a case the overall label on the vertical
axis will note this clipping, although it should be easy to see, because
the way it most often occurs is if the soft layers "bottom out" onto
the bone, which yields a short period of high stress, owing to the
high elastic modulus of bone. Each of the curves is filled in with a light
gray colour for stress/strength values up to 1, and with black for
higher values, making it easy to tell at a glance whether the
threat level is high.

\item \code{"whale acceleration"} for a time-series plot of whale acceleration.

\item \code{"blubber thickness"} for a time-series plot of blubber thickness.

\item \code{"sublayer thickness"} for a time-series plot of the thickness
of the layer interior to the blubber.

\item \code{"reactive forces"} for a time-series showing the reactive
forces associated with skin stretching (solid) and the compression of the
blubber and sublayer components (dashed).

\item \code{"compression stress"} for a time-series plot of the compression stress on the blubber
and the layer to its interior. (These stresses are equal, under an equilibrium assumption.)

\item \code{"skin stress"} for a time-series of skin stress in the along-skin y and z directions.

\item \code{"values"} for a listing of \code{param} values.

\item \code{"all"} for all of the above.

\item \code{"default"} for a three-element plot showing \code{"location"},
\code{"section"}, and \code{"threat"}.
}}

\item{drawEvents}{Logical, indicating whether to draw lines for some events,
such as the moment of closest approach.}

\item{colwcenter}{Colour used to indicate the whale centre.}

\item{colwinterface}{Colour used to indicate the interface
between whale blubber and sublayer.}

\item{colwskin}{Colour used to indicate the whale skin.}

\item{cols}{As \code{colw}, but the colour to be used for the ship bow location,
which is drawn with a dashed line.}

\item{colThreat}{Two-element colour specification used in \code{"threat"}.
The first colour is used for threat levels between 0 and 1, the second
for levels exceeding 1. The defaults are to use light gray and black.}

\item{lwd}{Line width used in plots for time intervals in which damage
criteria are not exceeded.}

\item{D}{Factor by which to thicken lines during times during which damage
criteria are exceeded.}

\item{debug}{Integer indicating debugging level, 0 for quiet operation and higher values
for more verbose monitoring of progress through the function.}

\item{...}{Ignored.}
}
\description{
Creates displays of various results of a simulation performed
with \code{\link{strike}}.
}
\examples{
## 1. default 3-panel plot
t <- seq(0, 0.7, length.out=200)
state <- c(xs=-2, vs=10*0.5144, xw=0, vw=0) # 10 knot ship
parms <- parameters(ms=20e3, lw=13)
sol <- strike(t, state, parms)
par(mar=c(3,3,1,1), mgp=c(2,0.7,0), mfrow=c(1, 3))
plot(sol)
## 2. all 12 plot types
par(mar=c(3,3,1,1) ,mgp=c(2,0.7,0), mfrow=c(4,3))
plot(sol, "all")
}
\references{
See \link{whalestrike} for a list of references.
}
