% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.strike}
\alias{plot.strike}
\title{Plot a strike object}
\usage{
\method{plot}{strike}(
  x,
  which = "default",
  drawEvents = TRUE,
  colwcenter = "black",
  colwinterface = "black",
  colwskin = "black",
  cols = "black",
  colThreat = c("white", "lightgray", "darkgray", "black"),
  lwd = 1,
  D = 3,
  debug = 0,
  ...
)
}
\arguments{
\item{x}{An object created by \code{\link[=strike]{strike()}}.}

\item{which}{A character vector that indicates what to plot.
This choices for its entries are listed below, in no particular order.
\itemize{

\item \code{"location"} for a time-series plot of boat location \code{xw} in
dashed black, whale centerline \code{xs} in solid gray,
blubber-interior interface in red, and skin in blue. The maximum
acceleration of ship and whale (in "g" units) are indicated in notes
placed near the horizontal axes. Those acceleration indications report
just a single value for each of ship and whale, but if the blubber
and sublayer have been squeezed to their limits, yielding a short and
intense force spike as the bone compresses, then the summaries will
also report on the spike duration and intensity. The spike is computed
based on using \code{\link[=runmed]{runmed()}} on the acceleration data, with a
\code{k} value that is set to correspond to 5 ms, or to k=11, whichever
is larger.

\item \code{"section"} to plot skin thickness, blubber thickness and sublayer thickness
in one panel, creating a cross-section diagram.

\item \code{"threat"} a stacked plot showing time-series traces of an
ad-hoc measure of the possible threat to skin, blubber and sublayer.
The threat level is computed as the ratio
of stress to ultimate strength, e.g. for blubber, it is
\code{x$WCF$stress/x$parms$s[2]}. The same vertical scale is used in
each of the subpanels that make up the stack. Any values exceeding
10 are clipped to 10, and in such a case the overall label on the vertical
axis will note this clipping, although it should be easy to see, because
the way it most often occurs is if the soft layers "bottom out" onto
the bone, which yields a short period of very high stress, owing to the
very high compression modulus of bone. Each of the curves is filled in with a light
gray colour for stress/strength values up to 1, and with black
for higher values; this makes it easy to tell at a glance whether the
threat level is high.

\item \code{"whale acceleration"} for a time-series plot of whale acceleration.

\item \code{"blubber thickness"} for a time-series plot of blubber thickness.

\item \code{"sublayer thickness"} for a time-series plot of the thickness
of the layer interior to the blubber.

\item \code{"reactive forces"} for a time-series showing the reactive
forces associated with skin stretching (solid) and the compression of the
blubber and sublayer components (dashed).

\item \code{"compression stress"} for a time-series plot of the compression stress on the blubber
and the layer to its interior. (These stresses are equal, under an equilibrium assumption.)

\item \code{"skin stress"} for a time-series of skin stress in the along-skin y and z directions.

\item \code{"lethality index"} for a time-series of Lethality Index, computed from compression stress
using \code{\link[=lethalityIndexFromStress]{lethalityIndexFromStress()}}. Values of Lethality Index that exceed 0.5 are highlighted,
and a dotted line is drawn at that value.

\item \code{"values"} for a listing of \code{param} values.

\item \code{"all"} for all of the above.

\item \code{"default"} for a three-element plot showing \code{"location"},
\code{"section"}, and \code{"threat"}.
}}

\item{drawEvents}{Logical, indicating whether to draw lines for some events,
such as the moment of closest approach.}

\item{colwcenter}{Colour used to indicate the whale centre.}

\item{colwinterface}{Colour used to indicate the interface
between whale blubber and sublayer.}

\item{colwskin}{Colour used to indicate the whale skin.}

\item{cols}{As \code{colw}, but the colour to be used for the ship bow location,
which is drawn with a dashed line.}

\item{colThreat}{a 4-element colour specification used in \code{"threat"} plots.
The colour transitions depend on the layer being plotted.
For skin and bone (i.e. in the bottom and top subpanels of the plot),
the first colour is used up to a stress:strength ratio
of 1/4, the second up to 1/2, the third up to 3/4, and the fourth above
3/4.  For blubber and sublayer (i.e. the second and third subpanels,
counting from the bottom), the colours are defined by comparing compressive
stress to the values of \code{tau25}, \code{tau50} and \code{tau75} (see the \code{\link[=parameters]{parameters()}}
documentation), with the first colour for stress under \code{tau25}, the second
for stress up to \code{tau50}, the third for stress up to \code{tau75}, and the fourth
for higher stresses.}

\item{lwd}{Line width used in plots for time intervals in which damage
criteria are not exceeded.}

\item{D}{Factor by which to thicken lines during times during which damage
criteria are exceeded.}

\item{debug}{Integer indicating debugging level, 0 for quiet operation and higher values
for more verbose monitoring of progress through the function.}

\item{...}{Ignored.}
}
\description{
Creates displays of various results of a simulation performed
with \code{\link[=strike]{strike()}}.
}
\examples{
# 1. default 3-panel plot
t <- seq(0, 0.7, length.out = 200)
state <- c(xs = -2, vs = knot2mps(12), xw = 0, vw = 0) # 12 knot ship
parms <- parameters() # default values
sol <- strike(t, state, parms)
par(mar = c(3, 3, 1, 1), mgp = c(2, 0.7, 0), mfrow = c(1, 3))
plot(sol)
# 2. all 12 plot types
par(mar = c(3, 3, 1, 1), mgp = c(2, 0.7, 0), mfrow = c(4, 3))
plot(sol, "all")

}
\references{
See \code{\link[=whalestrike]{whalestrike()}} for a list of references.
}
\author{
Dan Kelley
}
