% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ship.R
\name{shipAreaFromMass}
\alias{shipAreaFromMass}
\title{Compute ship wetted area from mass}
\usage{
shipAreaFromMass(ms)
}
\arguments{
\item{ms}{Ship mass (kg).}
}
\value{
Estimated area (m^2).
}
\description{
Estimate the wetted area of a Cape Islander boat,
given the vessel mass.
}
\details{
The method is based on scaling up the results for a single Cape
Islander ship, of displacement 20.46 tonnes, length 11.73m,
beam 4.63m, and draft 1.58m, on the assumption that the wetted area
is proportional to
\eqn{length*(2*draft+beam)}{length*(2*draft+beam)}.
This reference area is scaled to
the specified mass, \code{ms}, by multiplying by the 2/3
power of the mass ratio.

Note that this is a crude calculation meant as a stop-gap measure, for
estimates values of the \code{Ss} argument to \code{\link[=parameters]{parameters()}}.
It should not be used in preference to inferences
made from architectural drawings of a given ship under study.
}
\seealso{
Other functions relating to ship characteristics: 
\code{\link{shipLength}()},
\code{\link{shipMassFromLength}()},
\code{\link{shipWaterForce}()}
}
\author{
Dan Kelley
}
\concept{functions relating to ship characteristics}
