% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{shipWaterForce}
\alias{shipWaterForce}
\title{Ship water force}
\usage{
shipWaterForce(vs, parms)
}
\arguments{
\item{vs}{ship velocity [m/s]}

\item{parms}{A named list holding model parameters, as created by
\code{\link{parameters}}.}
}
\value{
Water drag force [N]
}
\description{
Estimate the retarding force of water on the ship, based on a drag law
\eqn{(1/2)*rho*CD*A*v^2}{(1/2)*rho*CD*A*v^2}
where \code{rho} is 1024 (kg/m^3), \code{CD} is \code{parms$CDs} and
\code{A} is \code{parms$as}.
}
\details{
This function is called by \code{\link{strike}}
with the initial ship velocity, to find an engine propulsive force that
is assumed to be constant throughout the impact.
}
