% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{skinForce}
\alias{skinForce}
\title{Skin force}
\usage{
skinForce(xs, xw, parms)
}
\arguments{
\item{xs}{Ship position [m]}

\item{xw}{Whale position [m]}

\item{parms}{Parameters of the simulation (as provided to \code{\link{simulate}}).}
}
\value{
Stretch-resisting skin force [N]
}
\description{
Skin force
}
\section{DEVELOPMENT NOTE}{
 think about formulation (re theta and B).
}

