% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{skinForce}
\alias{skinForce}
\title{Skin force}
\usage{
skinForce(xs, xw, parms)
}
\arguments{
\item{xs}{Ship position [m]}

\item{xw}{Whale position [m]}

\item{parms}{Parameters of the simulation (as provided to \code{\link{strike}}). The
quantities used by \code{skinForce} are \code{B} (ship beam width, m),
\code{D} (ship draft, in m), \code{Es} (skin elastic modulus in Pa),
\code{delta} (skin thickness in m), and \code{theta} (skin bevel angle
degrees, measured from a vector normal to undisturbed skin).}
}
\value{
A list containing \code{F}, the normal force [N], along with
\code{sigmay} and \code{sigmaz}, which are stresses [Pa] in the y (beam)
and z (draft) directions.
}
\description{
Skin force
}
