% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{stressFromStrainFunction}
\alias{stressFromStrainFunction}
\title{Create a function for stress in laminated layers}
\usage{
stressFromStrainFunction(l, a, b, N = 1000)
}
\arguments{
\item{l}{vector of layer thicknesses}

\item{a}{vector of multipliers}

\item{b}{vector of e-fold parameters}

\item{N}{integer specifying how many segments to use in the spline}
}
\value{
A spline function, created with \code{\link{splinefun}},
that returns stress as a function of total strain of the
system of compressing layers. For the purposes of the whale-strike
analysis, the strain should be between 0 and 1, i.e. there is
no notion of compressing blubber, etc. to negative thickess.
}
\description{
Assuming that unforced layer thicknesses are \eqn{l_i}, and that within each layer
the stress is given by
\deqn{a_i*[exp(b_i*\epsilon)-1]}
with strain \eqn{\epsilon}
being \eqn{\Delta l_i/l_i}, then the change in the
total thickness \eqn{L=\sum l_i} obeys
\deqn{0 = \Delta L - \sum (l_i /b_i) ln(1+\sigma / a_i)}
where \eqn{ln} is the natural logarithm. This formula rests on the assumption
that the stress, \eqn{\sigma}, is the same in each layer.
This expression is not easily inverted to get
\eqn{\sigma} in terms of \eqn{\Delta L}, but it may be solved
easily for particular numerical vaues, using \code{\link{uniroot}}.
This is done for a sequence of \code{N} values of strain \eqn{\epsilon}
that range from 0 to 1. Then a cubic spline is created to represent these
the relationship between \eqn{\sigma} and \eqn{\Delta L} values,
and this spline function is returned.
The purpose is to speed up processing of simulations carried out
with \code{\link{strike}}. In the authors tests, the speedup was by
an order of magnitude, and the loss of accuracy was
negligible, amounting to fractional errors in the 8th decimal place.
}
\examples{
library(whalestrike)
## Set blubber parameters for each layer, to see if
## we recover the raymond2007 data.
param <- parameters(a=rep(1.64e5,4), b=rep(2.47,4))
x <- seq(0, 0.5, length.out=100)
y <- param$stressFromStrain(x)
plot(x, y, type='l', lwd=4, col="gray")
data("raymond2007")
points(raymond2007$strain, raymond2007$stress, col=2)
}
