% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{strike}
\alias{strike}
\title{Simulate the collision of a ship and a whale}
\usage{
strike(t, state, parms, debug = 0)
}
\arguments{
\item{t}{time [s].}

\item{state}{A named vector holding the initial state of the model:
ship position \code{xs} [m],
ship speed \code{vs} [m/s],
whale position \code{xw} [m]
and whale speed \code{vw} [m/s].}

\item{parms}{A named list holding model parameters:
ship mass \code{"ms"} [kg],
ship beam \code{"B"} [m],
ship draft \code{"D"} [m],
whale mass \code{"mw"} [kg],
whale skin thickness \code{"delta"} [m],
whale skin elastic modulus \code{"Eskin"} [Pa],
whale skin deformation angle \code{"theta"} [deg],
whale blubber thickness \code{"beta"} [m]
and \code{"blubbermodel"} (passed to \code{\link{blubberForce}}.)}

\item{debug}{Integer indicating debugging level, 0 for quiet operation and higher values
for more verbose monitoring of progress through the function.}
}
\value{
An object of class \code{"whalestrike"}, consisting of a
list containing vectors for time (\code{t} [s]), ship position (\code{xs} [m]),
boat speed (\code{vs} [m/s]), whale position (\code{xw} [m]), whale speed (\code{vw} [m/s]),
boat acceleration (\code{dvsdt} [m/s^2]), and whale acceleration (\code{dvwdt} [m/s^2]),
along with a list containing the model parameters (\code{parms}).
}
\description{
Newtonian mechanics are used, taking the ship as undeformable,
and the whale as being cushioned by a skin layer and a blubber layer.
The forces are calculated by
\code{\link{skinForce}},
\code{\link{blubberForce}}, and
\code{\link{waterForce}}.
}
\examples{
library(whalestrike)
t <- seq(0, 1, length.out=500)
state <- c(xs=-1.5, vs=5, xw=0, vw=0)
parms <- list(ms=20e3, B=3, D=3,
              mw=20e3,
              delta=0.02, Eskin=2e7, theta=30,
              beta=0.3, blubbermodel="exponential")
sol <- strike(t, state, parms)
par(mfcol=c(3, 3), mar=c(2, 3, 1, 0.5), mgp=c(2, 0.7, 0), cex=0.7)
plot(sol, which="all")
}
