% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{strike}
\alias{strike}
\title{Simulate the collision of a ship and a whale}
\usage{
strike(t, state, parms, debug = 0)
}
\arguments{
\item{t}{a suggested vector of times [s] at which the simulated state will be reported.
This is only a suggestion, however, because \code{strike} is set up to detect high
accelerations caused by bone compression, and may set a finer reporting interval,
if such accelerations are detected. The detection is based on thickness of
compressed blubber and sublayer; if either gets to zero thickness, then
a new time grid is constructed, with 10 points during the timescale for
bone compression, which is assumed to be
\eqn{2*sqrt(Ly*Lz*a[4]*b[4]/(l[4]*mw)}, with terms as discussed in
the documentation for \code{\link{parameters}}. If this grid is finer
than the grid in the stated \code{t}, then the simulatoin is redone
using the new grid.}

\item{state}{A list or named vector holding the initial state of the model:
ship position \code{xs} [m],
ship speed \code{vs} [m/s],
whale position \code{xw} [m]
and whale speed \code{vw} [m/s].}

\item{parms}{A named list holding model parameters, created by
\code{\link{parameters}}.}

\item{debug}{Integer indicating debugging level, 0 for quiet operation and
higher values for more verbose monitoring of progress through the function.}
}
\value{
An object of class \code{"whalestrike"}, consisting of a
list containing vectors for time (\code{t} [s]), ship position (\code{xs} [m]),
boat speed (\code{vs} [m/s]), whale position (\code{xw} [m]), whale speed (\code{vw} [m/s]),
boat acceleration (\code{dvsdt} [m/s^2]), and whale acceleration (\code{dvwdt} [m/s^2]),
a list containing the model parameters (\code{parms}), a list with the results of
the skin force calculation (\code{SWF}), a list with the results of the compression
force calculations (\code{WCF}), and a vector of whale water force (\code{WWF}).
}
\description{
Newtonian mechanics are used, taking the ship as undeformable,
and the whale as being cushioned by a skin layer and a blubber layer.
The forces are calculated by
\code{\link{shipWaterForce}},
\code{\link{whaleSkinForce}},
\code{\link{whaleCompressionForce}}, and
\code{\link{whaleWaterForce}}.
}
\examples{
library(whalestrike)
t <- seq(0, 0.7, length.out=200)
state <- list(xs=-2, vs=knot2mps(10), xw=0, vw=0) # ship speed 10 knots
parms <- parameters()
sol <- strike(t, state, parms)
par(mfcol=c(1, 3), mar=c(3, 3, 0.5, 2), mgp=c(2, 0.7, 0), cex=0.7)
plot(sol)

}
\references{
See \link{whalestrike} for a list of references.
}
\author{
Dan Kelley
}
