% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{strike}
\alias{strike}
\title{Simulate the collision of a ship and a whale}
\usage{
strike(t, state, parms, debug = 0)
}
\arguments{
\item{t}{time [s].}

\item{state}{A named vector holding the initial state of the model:
ship position \code{xs} [m],
ship speed \code{vs} [m/s],
whale position \code{xw} [m]
and whale speed \code{vw} [m/s].}

\item{parms}{A named list holding model parameters, created by
\code{\link{parameters}}.}

\item{debug}{Integer indicating debugging level, 0 for quiet operation and
higher values for more verbose monitoring of progress through the function.}
}
\value{
An object of class \code{"whalestrike"}, consisting of a
list containing vectors for time (\code{t} [s]), ship position (\code{xs} [m]),
boat speed (\code{vs} [m/s]), whale position (\code{xw} [m]), whale speed (\code{vw} [m/s]),
boat acceleration (\code{dvsdt} [m/s^2]), and whale acceleration (\code{dvwdt} [m/s^2]),
along with a list containing the model parameters (\code{parms}).
}
\description{
Newtonian mechanics are used, taking the ship as undeformable,
and the whale as being cushioned by a skin layer and a blubber layer.
The forces are calculated by
\code{\link{shipWaterForce}},
\code{\link{whaleSkinForce}},
\code{\link{whaleCompressionForce}}, and
\code{\link{whaleWaterForce}}.
}
\examples{
library(whalestrike)
t <- seq(0, 1, length.out=500)
state <- c(xs=-2, vs=5, xw=0, vw=0)
ls <- 10           # ship length [m]
draft <- 1.5       # ship draft [m]
beam <- 3          # ship beam [m]
lw <- 11           # whale length [m]
impactWidth <- 0.5 # impact region width [m]
impactHeight <- 1  # impact region height [m]
parms <- parameters(ms=20e3, Ss=ls*(2*draft+beam),
              impactWidth=impactWidth, impactHeight=impactHeight,
              lw=lw, mw=whaleMassFromLength(lw),
              Sw=whaleAreaFromLength(lw, "wetted"),
              alpha=0.02, Ealpha=2e7, theta=45,
              beta=0.3, Ebeta=6e5,
              gamma=0.5, Egamma=4e5)
sol <- strike(t, state, parms)
par(mfcol=c(2, 3), mar=c(2, 3, 1, 0.5), mgp=c(2, 0.7, 0), cex=0.7)
plot(sol)

}
\references{
See \link{whalestrike} for a list of references.
}
