% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{summary.parameters}
\alias{summary.parameters}
\title{Summarize a parameters object}
\usage{
\method{summary}{parameters}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"parameters"}, as created with \code{\link[=parameters]{parameters()}}.}

\item{\dots}{ignored}
}
\value{
\code{summary.parameters} returns nothing. It is called for
its side effect of printing information about the parameters
in a ship-whale collision simulation.
}
\description{
This provides an overview of the contents of an object
created with \code{\link[=parameters]{parameters()}}.
}
\examples{
summary(parameters())

}
\author{
Dan Kelley
}
