% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{summary.strike}
\alias{summary.strike}
\title{Summarize the parameters of a simulation, and its results}
\usage{
\method{summary}{strike}(object)
}
\arguments{
\item{object}{An object created by \code{\link[=strike]{strike()}}.}
}
\description{
Summarize the parameters of a simulation, and its results
}
\references{
See \code{\link[=whalestrike]{whalestrike()}} for a list of references.
}
\author{
Dan Kelley
}
