% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\docType{data}
\name{versionOfDefaults}
\alias{versionOfDefaults}
\title{Version of parameter defaults}
\description{
A Either "2018" (to get default parameter values from the work
in summer of 2018) or "2019a" to get parameter values as of
the start of summer, 2019.
}
