% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{whale_measurements}
\alias{whale_measurements}
\title{Get values for various whale measurements}
\usage{
whale_measurements(species = NULL)
}
\arguments{
\item{species}{either (1) the name of a species or (2) NULL. In the first
case, the table is consulted to find a row with the given species
name, and that row is returned. In the second case, the whole table
is returned.}
}
\value{
The return value contains
\itemize{
\item \code{name} species name, as used in e.g. whaleMassFromLength().
\item \code{Species} proper species name. (This is not used in this package.)
\item \code{length} whale length in metres.
\item \code{bone} whale bone thickness in metres, measured from the centre to the sublayer.
\item \code{sublayer} thickness of sublayer in meters; this was called \code{muscle} in Mayette's document.
\item \code{blubber} whale blubber thickness in meters.
\item \code{skin} whale skin thickness in metres.
}
}
\description{
This uses a data frame containing information about several whale
species, compiled by Alexandra Mayette and provided to Dan
Kelley as a personal communication on 2026-01-28.
}
\details{
There are two species in the table that are not in Mayette's table.
These are \code{"Pac. Right Whale"} and \code{"Bryde Whale"}. For these, Mayette has suggesting
using values for the \code{"N. Atl. Right Whale"} and \code{"Sei Whale"} cases, respectively,
as conditional estimates for use in this package.
}
\author{
Dan Kelley, using data and advice from Alexandra Mayette
}
