% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whale_measurements}
\alias{whale_measurements}
\title{Whale measurements}
\description{
\code{whale_measurements} is a data frame containing information about several whale
species, compiled by Alexandra Mayette and provided to Dan
Kelley as a personal communication on 2026-01-28.
}
\details{
The columns are follows.
\itemize{
\item \code{name} species name, as used in this package.
\item \code{Species} proper species name, not used in this package.
\item \code{length} whale length in metres.
\item \code{girth} whale girth in centimetres.
\item \code{bone} whale bone thickess in centimeters.
\item \code{sublayer} thickness of sublayer in centimeters; this was called muscle in the Mayette document
\item \code{blubber} whale blubber thickness in centimeters.
\item \code{skin} whale skin thicness in centimetres.
}
}
\examples{
library(whalestrike)
data(whale_measurements)
whale_measurements

}
