% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleAreaFromLength}
\alias{whaleAreaFromLength}
\title{Whale projected area, as function of length}
\usage{
whaleAreaFromLength(L, type = "wetted")
}
\arguments{
\item{L}{length in m}

\item{type}{character string indicating the type of area, with
\code{"projected"} for side-projected area using 0.143L^2,
and \code{"surface"}
for submerged surface wetted, calculated by spinning
the necropsiy side-view presented in Daoust et al. (2018)
along the animal axis, yielding 0.737L^2. The \code{"surface"}
version is suitable for use in \code{\link{whaleWaterForce}}.}
}
\description{
Whale projected area, as function of length
}
\references{
R worksheet \code{dek/20180623_whale_area.Rmd}, available
upon request.
}
