% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleAreaFromLength}
\alias{whaleAreaFromLength}
\title{Whale projected area, as function of length}
\usage{
whaleAreaFromLength(L, species = "N. Atl. Right Whale", type = "wetted")
}
\arguments{
\item{L}{whale length in metres.}

\item{species}{a string indicating the whale species. In
the present version of the package, this parameter is ignored,
and it is assumed that the formula developed for North
Atlantic Right Whales will be applicable to other species.
This is not a large concern, because the area only affects
the water drag, which will not be large during the
short interval of a ship impact.}

\item{type}{character string indicating the type of area, with
\code{"projected"} for a side-projected area, and
\code{"wetted"} for the total wetted area. The wetted
area was computed by mathematically spinning a spline fit to the
side-view. In both cases, the original data source is the
necropsy side-view presented in Daoust et al. (2018).}
}
\description{
This depends on calculations based on the digitized shape of
a whale necropsy, which is provided as \link{whale_shape}.
The results are
\eqn{0.143 * L^2}{0.143 * L^2}
for the projected area (see reference 1)
and
\eqn{0.448 * (0.877 * L)^2}{0.448 * (0.877 * L)^2}
for the wetted area
(see reference 2, but note that we use a correction related to whale mass).
}
\details{
Note that multiple digitizations were done,
and that the coefficients used in the formulae
agreed to under 0.7 percent percent between these digitizations.
}
\examples{
L <- 3:20
A <- whaleAreaFromLength(L)
plot(L, A, xlab="Length [m]", ylab="Area [m^2]", type="l")

}
\references{
\enumerate{
\item Dan Kelley's internal document \verb{dek/20180623_whale_area.Rmd}, available
upon request.
\item Dan Kelley's internal document \verb{dek/20180707_whale_mass.Rmd}, available
upon request.
\item Daoust, Pierre-Yves, Emilie L. Couture, Tonya Wimmer, and Laura Bourque.
"Incident Report. North Atlantic Right Whale Mortality Event in the Gulf of St.
Lawrence, 2017." Canadian Wildlife Health Cooperative, Marine Animal Response
Society, and Fisheries and Oceans Canada, 2018.
\url{https://publications.gc.ca/site/eng/9.850838/publication.html}.
}
}
\author{
Dan Kelley
}
