% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleAreaFromLength}
\alias{whaleAreaFromLength}
\title{Whale projected area, as function of length}
\usage{
whaleAreaFromLength(L, type = "wetted")
}
\arguments{
\item{L}{length in m}

\item{type}{character string indicating the type of area, with
\code{"projected"} for side-projected area using 0.143L^2,
and \code{"wetted"}
for submerged surface wetted, calculated by spinning
the necropsy side-view presented in Daoust et al. (2018)
along the animal axis, yielding
0.451 * (0.8715 * L)^2, where the direct factor on L
was developed by comparing similarly-predicted masses to
the results of \code{\link{whaleMassFromLength}}, as described
in [2].}
}
\description{
Whale projected area, as function of length
}
\references{
1. Dan Kelley's internal document \code{dek/20180623_whale_area.Rmd}, available
upon request.

2. Dan Kelley's internal document \code{dek/20180707_whale_mass.Rmd}, available
upon request.
}
