% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleBlubberForce}
\alias{whaleBlubberForce}
\title{Blubber force}
\usage{
whaleBlubberForce(xs, xw, parms)
}
\arguments{
\item{xs}{Ship position [m]}

\item{xw}{Whale position [m]}

\item{parms}{A named list holding model parameters, as created by
\code{\link{parameters}}.}
}
\value{
Compression-resisting force [N]
}
\description{
Calculate the reaction force of blubber, as the product of stress
and area. Stress is computed as the product of blubber elastic
modulus \code{parms$Eb} times strain, where the latter is computed
from ship-whale separation using \code{1-(xw-xs)/parms$beta} if
\code{xw} is between \code{xs} and \code{xs+parms$beta},
or zero otherwise. Area is computed as the product of
\code{parms$B} and \code{parms$D}.
}
\references{
See \link{whalestrike} for a list of references.
}
