% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleCompressionForce}
\alias{whaleCompressionForce}
\title{Whale compression force}
\usage{
whaleCompressionForce(xs, xw, parms)
}
\arguments{
\item{xs}{Ship position [m]}

\item{xw}{Whale position [m]}

\item{parms}{A named list holding model parameters, created by
\code{\link{parameters}}.}
}
\value{
A list containing \code{force} [N], the
compression-resisting force, \code{stress} [Pa], the ratio
of that force to the impact area, \code{strain}, the total
strain, \code{alphaCompressed} [m],
the skin thickness, \code{betaCompressed} [m],
the blubber thickness, and \code{gammaCompressed} [m],
the sublayer thickness.
}
\description{
Calculate the reaction force of blubber and the sublayer to
its interior, as the product of stress and area.
Linear compression dynamics is assumed for each layer, i.e. that
stress is strain times elastic modulus. Static dynamics is assumed,
so that the stresses match in the the two layers. The calculation
is done in a two-step process. First, the stress for an equivalent
layer is computed, based on a combined elastic modulus. If this
stress exceeds the blubber modulus, the blubber thickness is set to zero
and the calculation is redone with just the sublayer. If not,
the layer thickness are computed assuming equal stress in each layer.
Force is calculated as the product of stress and the area
given by the product of \code{parms$impactWidth} and \code{parms$impactHeight}.
}
\references{
See \link{whalestrike} for a list of references.
}
