% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forces.R
\name{whaleCompressionForce}
\alias{whaleCompressionForce}
\title{Whale compression force}
\usage{
whaleCompressionForce(xs, xw, parms)
}
\arguments{
\item{xs}{Ship position (m).}

\item{xw}{Whale position (m).}

\item{parms}{A named list holding model parameters, created by
\code{\link[=parameters]{parameters()}}.}
}
\value{
A list containing: \code{force} (N), the
compression-resisting force; \code{stress} (Pa), the ratio
of that force to the impact area; \code{strain}, the total
strain, and \code{compressed}, a four-column matrix (m)
with first column for skin compression, second for blubber
compression, third for sublayer compression, and fourth
for bone compression.
}
\description{
Calculate the total compression stress and force, along
with the thicknesses of skin, blubber, sublayer, and bone.
The stress is computed with the \code{\link[=stressFromStrainFunction]{stressFromStrainFunction()}} function that
is created by \code{\link[=parameters]{parameters()}} and stored in \code{para}.
the force is computed by multiplying stess by area
computed as the product of \code{parms$Ly} and \code{parms$Lz}.
Any negative layer thicknesses are set to zero, as a way to
avoid problems with aphysical engineering compression strains that
exceed 1.
}
\references{
See \code{\link[=whalestrike]{whalestrike()}} for a list of references.
}
\seealso{
Other functions relating to whale characteristics: 
\code{\link{stressFromStrainFunction}()},
\code{\link{whaleLengthFromMass}()},
\code{\link{whaleMassFromLength}()},
\code{\link{whaleSkinForce}()},
\code{\link{whaleWaterForce}()}

Other functions relating to forces: 
\code{\link{shipWaterForce}()},
\code{\link{stressFromStrainFunction}()},
\code{\link{whaleSkinForce}()},
\code{\link{whaleWaterForce}()}
}
\author{
Dan Kelley
}
\concept{functions relating to forces}
\concept{functions relating to whale characteristics}
