% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleLengthFromMass}
\alias{whaleLengthFromMass}
\title{Compute whale length from mass}
\usage{
whaleLengthFromMass(M, species = "N. Atl. Right Whale", model = "fortune2012")
}
\arguments{
\item{M}{Whale mass (kg).}

\item{species}{A string indicating the whale species
(see \code{\link[=whaleMassFromLength]{whaleMassFromLength()}} for details).}

\item{model}{Character string specifying the model
(see \code{\link[=whaleMassFromLength]{whaleMassFromLength()}} for details).}
}
\value{
Whale length (m).
}
\description{
This works by inverting \code{\link[=whaleMassFromLength]{whaleMassFromLength()}} using \code{\link[=uniroot]{uniroot()}}.
}
\references{
See \code{\link[=whalestrike]{whalestrike()}} for a list of references.
}
\seealso{
\code{\link[=whaleMassFromLength]{whaleMassFromLength()}} is the reverse of this.
}
\author{
Dan Kelley
}
