% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_geometry.R
\name{whaleMassFromLength}
\alias{whaleMassFromLength}
\title{Whale mass inferred from length}
\usage{
whaleMassFromLength(L, species = "N. Atl. Right Whale", model = NULL)
}
\arguments{
\item{L}{whale length in m.}

\item{species}{character value specifying the species (see \dQuote{Details}).
If  only one value is given, then it will repeated to have the same length
as \code{L}. Otherwise, its length must match the length of \code{L}.}

\item{model}{either NULL (the default), to choose a model based on the
particular species, or a character value specifying the model (see \dQuote{Details}).
If  only one value is given, then it will repeated to have the same length
as \code{L}. Otherwise, its length must match the length of \code{L}.}
}
\value{
Mass in kg.
}
\description{
Calculate an estimate of the mass of different species of whale,
based on animal length, based on formulae as listed in
\dQuote{Details}.
}
\details{
The permitted values for \code{model} and \code{species} are as follows. Note
that if \code{model} is not provided, then Fortune (2012) is used
for \code{"N. Atl. Right Whale"}, and Lockyer (1976) is used for all
other species.
\itemize{
\item \code{"moore2005"} (which only works if \code{species} is \code{"N. Atl. Right Whale"}) yields
\eqn{242.988 * exp(0.4 * length)}{242.988 * exp(0.4 * L)},
which (apart from a unit change on \code{L}) is the regression equation
shown above Figure 1d in Moore et al. (2005) for right whales. A
difficult in the Moore et al. (2005) use of a single nonzero digit
in the multiplier on \code{L} is illustrated in \dQuote{Examples}.
\item \code{"fortune2012"} with \code{species="N. Atl. Right Whale"} yields the formula
\eqn{exp(-10.095 + 2.825*log(100*L))}{exp(-10.095 + 2.825*log(100*L))}
for North Atlantic right whales, according to a corrected version of the
erroneous formula given in the caption of Figure 4 in Fortune et al (2012).
(The error, an exchange of slope and intercept, was confirmed by
S. Fortune in an email to D. Kelley dated June 22, 2018.)
\item \code{"fortune2012"} with \code{species="N. Pac. Right Whale"} yields the formula
\eqn{exp(-12.286 + 3.158*log(100*L))}{exp(-12.286 + 3.158*log(100*L))}
for North Pacific right whales, according to a corrected version of the
erroneous formula given in the caption of Figure 4 in Fortune et al (2012).
(The error, an exchange of slope and intercept, was confirmed by
S. Fortune in an email to D. Kelley dated June 22, 2018.)
\item \code{"lockyer1976"} uses formulae from Table 1 of Lockyer (1976). The
permitted \code{species} and the formulae used are as follows (note that
the \code{"Gray Whale"} formula is in the table's caption, not in the table itself).
\itemize{
\item \code{"Blue Whale"}:       \eqn{2.899 L^{3.25}}{2.899 * L^3.25}
\item \code{"Bryde Whale"}:      \eqn{12.965 L^{2.74}}{12.965 * L^2.74}
\item \code{"Fin Whale"}:        \eqn{7.996 L^{2.90}}{7.996 * L^2.90}
\item \code{"Gray Whale"}:       \eqn{5.4 L^{3.28}}{5.4 * L^3.28}
\item \code{"Humpback Whale"}:   \eqn{16.473 L^{2.95}}{16.473 * L^2.95}
\item \code{"Minke Whale"}:      \eqn{49.574 L^{2.31}}{49.574 * L^2.31}
\item \code{"Pac. Right Whale"}: \eqn{13.200 L^{3.06}}{13.200 * L^3.06}
\item \code{"Sei Whale"}:        \eqn{25.763 L^{2.43}}{25.763 * L^2.43}
\item \code{"Sperm Whale"}:      \eqn{6.648 L^{3.18}}{6.648 * L^3.18}
}
}
}
\examples{
library(whalestrike)
L <- seq(5, 15, length.out = 100)
kpt <- 1000 # kg per tonne
# Demonstrate (with dashing) the sensitivity involved in the single-digit
# parameter in Moore's formula, and (with colour) the difference to the
# Fortune et al. (2012) formulae.
plot(L, whaleMassFromLength(L, model = "moore2005") / kpt,
    type = "l", lwd = 2,
    xlab = "Right-whale Length [m]", ylab = "Mass [tonne]"
)
lines(L, 242.988 * exp(0.35 * L) / kpt, lty = "dotted", lwd = 2)
lines(L, 242.988 * exp(0.45 * L) / kpt, lty = "dashed", lwd = 2)
lines(L, whaleMassFromLength(L,
    species = "N. Atl. Right Whale",
    model = "fortune2012"
) / kpt, col = 2, lwd = 2)
lines(L, whaleMassFromLength(L,
    species = "N. Pac. Right Whale",
    model = "fortune2012"
) / kpt, col = 3, lwd = 2)
grid()
legend("topleft",
    bg = "white", y.intersp = 1.4, lwd = 2, col = 1:3,
    legend = c("moore2005", "fortune2012 Atlantic", "fortune2012 Pacific")
)

# Emulate Figure 1 of Lockyer (1976), with roughly-chosen plot limits.
L <- seq(0, 18, 0.5)
m <- whaleMassFromLength(L, species = "Pac. Right Whale", model = "lockyer1976") / kpt
plot(L, m,
    col = 1, xlab = "Length [m]", ylab = "Mass [tonne]", type = "l", lwd = 2,
    xaxs = "i", yaxs = "i", xlim = c(3, 30), ylim = c(0, 180)
)
L <- seq(0, 28, 0.5)
m <- whaleMassFromLength(L, species = "Blue Whale", model = "lockyer1976") / kpt
lines(L, m, col = 2, lwd = 2)
L <- seq(0, 24, 0.5)
m <- whaleMassFromLength(L, species = "Fin Whale", model = "lockyer1976") / kpt
lines(L, m, col = 3, lwd = 2)
L <- seq(0, 18, 0.5)
m <- whaleMassFromLength(L, species = "Sei Whale", model = "lockyer1976") / kpt
lines(L, m, col = 1, lty = 2, lwd = 2)
L <- seq(0, 17, 0.5)
m <- whaleMassFromLength(L, species = "Bryde Whale", model = "lockyer1976") / kpt
lines(L, m, col = 2, lty = 2, lwd = 2)
L <- seq(0, 12, 0.5)
m <- whaleMassFromLength(L, species = "Minke Whale", model = "lockyer1976") / kpt
lines(L, m, col = 3, lty = 2, lwd = 2)
L <- seq(0, 17, 0.5)
m <- whaleMassFromLength(L, species = "Humpback Whale", model = "lockyer1976") / kpt
lines(L, m, col = 1, lty = 3, lwd = 2)
L <- seq(0, 18, 0.5)
m <- whaleMassFromLength(L, species = "Sperm Whale", model = "lockyer1976") / kpt
lines(L, m, col = 2, lty = 3, lwd = 2)
L <- seq(0, 15, 0.5)
m <- whaleMassFromLength(L, species = "Gray Whale", model = "lockyer1976") / kpt
lines(L, m, col = 3, lty = 3, lwd = 2)
grid()
legend("topleft",
    bg = "white", y.intersp = 1.4, col = c(1:3, 1:3, 1:2),
    lwd = 2, lty = c(rep(1, 3), rep(2, 3), rep(3, 3)),
    legend = c("Right", "Blue", "Fin", "Sei", "Bryde", "Minke", "Humpback", "Sperm", "Gray")
)

}
\references{
\itemize{
\item Lockyer, C. "Body Weights of Some Species of Large Whales." J. Cons. Int.
Explor. Mer. 36, no. 3 (1976): 259-73.
\item Moore, M.J., A.R. Knowlton, S.D. Kraus, W.A. McLellan, and R.K. Bonde.
"Morphometry, Gross Morphology and Available Histopathology in North Atlantic
Right Whale (Eubalaena Glacialis) Mortalities (1970 to 2002)." Journal of
Cetacean Research and Management 6, no. 3 (2005): 199-214.
\item Fortune, Sarah M. E., Andrew W. Trites, Wayne L. Perryman, Michael J. Moore,
Heather M. Pettis, and Morgan S. Lynn. "Growth and Rapid Early Development of
North Atlantic Right Whales (Eubalaena Glacialis)." Journal of Mammalogy 93,
no. 5 (2012): 1342-54. \doi{10.1644/11-MAMM-A-297.1}.
}
}
\seealso{
\code{\link[=whaleLengthFromMass]{whaleLengthFromMass()}} is the reverse of this.
}
\author{
Dan Kelley
}
