% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleSkinForce}
\alias{whaleSkinForce}
\title{Skin force}
\usage{
whaleSkinForce(xs, xw, parms)
}
\arguments{
\item{xs}{Ship position [m]}

\item{xw}{Whale position [m]}

\item{parms}{A named list holding model parameters, created by
\code{\link{parameters}}.}
}
\value{
A list containing \code{force}, the normal force [N], along with
\code{sigmay} and \code{sigmaz}, which are stresses [Pa] in the y (beam)
and z (draft) directions.
}
\description{
The ship-whale separation is used to calculate the deformation of the skin. The
parameters of the calculation are \code{parms$B} (impact area width, m),
\code{parms$D} (impact area height, in m), \code{parms$Eskin} (skin elastic modulus in Pa),
\code{parms$delta} (skin thickness in m), and \code{parms$theta} (skin bevel angle
degrees, measured from a vector normal to undisturbed skin).
}
\references{
See \link{whalestrike} for a list of references.
}
