% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whalestrike}
\alias{whalestrike}
\title{Perform dynamical simulation}
\usage{
whalestrike(t, state, parms)
}
\arguments{
\item{t}{time [s].}

\item{state}{model state, a vector containing boat position \code{X} [m],
boat speed \code{V} [m/s], whale position \code{x} [m],
and whale speed \code{v} [m/s].}

\item{parms}{a named vector holding model parameters.}
}
\value{
An object of class \code{"whalestrike"}, consisting of a
list  containing vectors for time (\code{t} [s]), boat position (\code{X} [m]),
boat speed (\code{V} [m/s]), whale position (\code{x} [m]), whale speed (\code{v} [m/s]),
boat acceleration (\code{dVdt} [m/s^2]), and whale acceleration (\code{dvdt} [m/s^2]),
along with a list containing the model parameters (\code{parms}).
}
\description{
Perform dynamical simulation
}
\examples{
t <- seq(0, 1, length.out=500)
state <- c(X=-1.5, V=5, x=0, v=0)
parms <- list(shipMass=20e3, beam=3, draft=3, V=4.6296,
              whaleMass=20e3, B=0.35, gamma=1, delta=0.02, ES=2e+07, blubbermodel="exponential")
sol <- whalestrike(t, state, parms)
plot(sol)
}
