% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whalestrike.R
\name{whaleWaterForce}
\alias{whaleWaterForce}
\title{Whale force}
\usage{
whaleWaterForce(vw, parms)
}
\arguments{
\item{vw}{Whale velocity [m/s]}

\item{parms}{A named list holding model parameters, as created by
\code{\link{parameters}}.}
}
\value{
Water drag force [N]
}
\description{
Estimate the retarding force of water on the whale, based on a drag law
\eqn{(1/2)*rho*CD*A*v^2}{(1/2)*rho*CD*A*v^2}
where \code{rho} is 1024 (kg/m^3), \code{CD} is \code{parms$CDw} and
\code{A} is \code{parms$aw}.
}
